Anoma.Block (Anoma v0.11.0)

I represent a block that will be stored and gossiped around the network.

Fields

:id - Identification number derived from the serialization of my:

digest(block)

round

pub_key

:block - The block I contain

:round - The round I come from

:pub_key - Public key

:signature - Id signed with the private_key related to my public_key

Types

private_key()

@type private_key() :: [:crypto.key_id()]

public_key()

@type public_key() :: [:crypto.key_id()]

t()

@type t() :: %Anoma.Block{
  block: Anoma.Block.Base.t(),
  id: binary(),
  pub_key: Anoma.Serializer.public_key(),
  round: non_neg_integer(),
  signature: binary() | nil
}

Functions

create(block, pub, round)

@spec create(
  Anoma.Block.Base.t(),
  Anoma.Serializer.public_key()
  | {Anoma.Serializer.public_key(), Anoma.Serializer.private_key()},
  non_neg_integer()
) :: t()

I create a block, if a private key is passed in my second parameter, then I also contain a signature of myself.

The signature (if created), represents that my :pub_key signed my :id. This can be cryptography verified with Serializer.verify

Parameters

block - the base block that I am structured around

key - Either a public key or a public private key paring

round - The round that I represent

create_table()

I create a :mnesia table for Anoma.Block. This table is backed by rocksdb, and thus persists across IEX sessions.

I will only ever needed to be called once upon Configuration start, Anoma.Mnesia.init/0 will likely set me up as is.

create_table(table_key, rocks?)

I am like create_table/0, however I am given a special table_key. This overrides the default table key of Anoma.Block.

I am useful when trying to spawn many solvers/validators/etc, who all want their own tables.

Parameters

table_key - the name of the table

rocks? - should we persist as a rocksdb table?

decode(arg)

encode(block)

@spec encode(t()) :: tuple()

encode(block, atom)

@spec encode(t(), atom()) :: tuple()

sign(block, priv)

@spec sign(t(), Anoma.Serializer.private_key()) :: binary()

I sign my id given a private key. I do no validation checking if the id is derived properly.