Anoma.Configuration (Anoma v0.19.1)

I am a configuration module. I read the provided TOML configuration file and feed the apporpriate info for Node launching

The codebase has a corresponding file that can inform the user of the format I expect.

Public API

configuration/0

configuration/1

read_configuration/1

node_settings/1

save/1

save/2

locate_dump_file/1

launch_min/1

launch_min/2

Types

configruation()

@type configruation() :: [{String.t(), section_format()}]

configuration_map()

@type configuration_map() :: %{required(String.t()) => section_map()}

launch_option()

@type launch_option() ::
  {:use_rocksdb, boolean()}
  | {:supervisor, [Supervisor.option()]}
  | {:testing, boolean()}

I control options for launch_min/2

###Options

:use_rocksdb - see Anoma.Node.configuration/0 for more information

:supervisor - This flag determine if we use a supervisor and if so what options. See Supervisor.option/0 for supervisor options

:testing - This flag notes if we are testing the node. This gets fed directly into the type Anoma.Node.configuration/0 for Anoma.Node.start_link/1. Please consult the Anoma.Node.configuration/0 documentation for the full effect this has on the node

node_settings()

@type node_settings() :: Keyword.t()

section_format()

@type section_format() :: [{String.t(), (term() -> boolean()), term()}]

section_map()

@type section_map() :: %{required(String.t()) => term()}

Functions

configuration(user_map \\ %{})

@spec configuration(configuration_map()) :: configuration_map()

configuration_format()

@spec configuration_format() :: configruation()

default_configuration_location(env \\ Mix.env())

default_data_location(env \\ Mix.env())

is_pinger(field)

@spec is_pinger(any()) :: boolean()

launch_min(parsed_map, options \\ [])

@spec launch_min(configuration_map(), [launch_option()]) :: GenServer.on_start()

Given a parsed map with minimal node startup info I launch the node with the appopriate name

Options

see launch_option/0 for the full list of optional arguments

locate_dump_file(path)

@spec locate_dump_file(configuration_map()) :: Path.t() | nil

node_settings(configuration)

@spec node_settings(configuration_map()) :: Anoma.Node.min_engine_configuration()

Given a map, I decode all the needed info for a minimal node startup and put it in the appropriate keyword list

read_configuration(file)

@spec read_configuration(Path.t()) :: configuration_map()

save(config, file \\ default_configuration_location())

@spec save(configuration_map(), Path.t()) :: Path.t()