Anoma.Crypto.Symmetric (Anoma v0.28.0)

Types

t()

@type t() :: xchacha()

I represent the symmetric types available to the system

xchacha()

@type xchacha() :: {:xchacha, xchacha_key(), xchacha_nonce()}

xchacha_key()

@type xchacha_key() :: <<_::256>>

xchacha_nonce()

@type xchacha_nonce() :: <<_::192>>

Functions

decrypt(binary, sym)

@spec decrypt(binary(), t()) :: term() | {:error, any()}

decrypt_raw(binary, arg)

@spec decrypt_raw(binary(), t()) :: binary() | {:error, any()}

encrypt(message, sym)

@spec encrypt(any(), t()) :: binary()

I encrypt data given any known schema and a message.

The message will be turned into binary via :erlang.term_to_binary, so please do not turn it to binary before hand.

encrypt_raw(binary, arg)

@spec encrypt_raw(binary(), t()) :: binary()

I encrypt data given any known schema and a message.

I am raw in that I do not serialize the data, Only use me if you know what you are doing.

random_xchacha()

@spec random_xchacha() :: xchacha()

random_xchacha_key()

@spec random_xchacha_key() :: xchacha_key()

random_xchacha_nonce()

@spec random_xchacha_nonce() :: xchacha_nonce()