Anoma.Dump (Anoma v0.17.0)

I provide an interface to dump current state and load appropriate external files to launch them as Anoma nodes.

You can also use me to dump info such as current states and tables in a readable map format as well as get info stored in external files in binary format.

Dumping API

I give access to following public dumping functionality:

dump/2

get_all/1

get_state/1

get_tables/1

Loading API

I give access to following public loading functionality

launch/2

launch/3

load/1

Types

clock_eng()

@type clock_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Clock.t()}

configuration_eng()

@type configuration_eng() ::
  {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Configuration.t()}

dump()

@type dump() :: %{
  router: Anoma.Crypto.Id.t(),
  transport: Anoma.Crypto.Id.t(),
  router_state: Anoma.Node.Router.t(),
  transport_id: Anoma.Crypto.Id.Extern.t(),
  logger_topic: Anoma.Crypto.Id.Extern.t(),
  mempool_topic: Anoma.Crypto.Id.Extern.t(),
  executor_topic: Anoma.Crypto.Id.Extern.t(),
  storage_topic: Anoma.Crypto.Id.Extern.t(),
  configuration: configuration_eng(),
  logger: log_eng(),
  clock: clock_eng(),
  ordering: ord_eng(),
  mempool: mem_eng(),
  pinger: ping_eng(),
  executor: ex_eng(),
  storage: storage_eng(),
  dumper: dump_eng(),
  storage_data: stores(),
  qualified: list(),
  order: list(),
  block_storage: list(),
  use_rocks: boolean()
}

dump_eng()

@type dump_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Dumper.t()}

ex_eng()

@type ex_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Executor.t()}

launch_option()

@type launch_option() :: {:supervisor, [Supervisor.option()]} | {:testing, boolean()}

I control launch options for launch/2

Options

:supervisor - This flag determine if we use a supervisor and if so what options. See Supervisor.option/0 for supervisor options

:testing - This flag notes if we are testing the node. This gets fed directly into the type Anoma.Node.configuration/0 for Anoma.Node.start_link/1. Please consult the Anoma.Node.configuration/0 documentation for the full effect this has on the node

log_eng()

@type log_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Logger.t()}

mem_eng()

@type mem_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Mempool.t()}

ord_eng()

@type ord_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Ordering.t()}

ping_eng()

@type ping_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Pinger.t()}

storage_eng()

@type storage_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Storage.t()}

stores()

@type stores() :: {Anoma.Node.Storage.t(), atom()}

Functions

dump(name, node)

@spec dump(Path.t(), atom()) :: {:ok, :ok} | {:error, any()}

I dump the current state with storage. I accept a string as a name, so that the resulting file will be created as name.txt in the appropriate data directory. As a second argument I accept a node name whose info presented as a map I dump as a binary.

Note that if the environment is test we do not use the XDG format for storing data and instead dump the files in the immadiate app folder.

The map typing can be seen in get_all

dump_full_path(name, node)

get_all(node)

@spec get_all(atom()) :: dump()

I get all the info on the node tables and engines in order:

router

logger

clock

ordering

mempool

pinger

executor

table names

qualified

order

block_storage And turn the info into a tuple

get_state(node)

@spec get_state(atom()) :: %{
  router: Anoma.Crypto.Id.t(),
  transport: Anoma.Crypto.Id.t(),
  router_state: Anoma.Node.Router.t(),
  transport_id: Anoma.Crypto.Id.Extern.t(),
  logger_topic: Anoma.Crypto.Id.Extern.t(),
  mempool_topic: Anoma.Crypto.Id.Extern.t(),
  executor_topic: Anoma.Crypto.Id.Extern.t(),
  storage_topic: Anoma.Crypto.Id.Extern.t(),
  configuration: configuration_eng(),
  logger: log_eng(),
  clock: clock_eng(),
  ordering: ord_eng(),
  mempool: mem_eng(),
  pinger: ping_eng(),
  executor: ex_eng(),
  storage: storage_eng(),
  dumper: dump_eng()
}

I get the engine states in order:

router

mempool topic

executor topic

dumper

storage

logger

clock

ordering

mempool

pinger

executor

get_tables(node)

@spec get_tables(atom()) :: %{
  storage_data: stores(),
  qualified: list(),
  order: list(),
  block_storage: list(),
  use_rocks: boolean()
}

I get the node tables in order:

storage (names)

qualified

order

block_storage

launch(file, name, options \\ [])

@spec launch(String.t(), atom(), [launch_option()]) ::
  {:ok,
   %Anoma.Node{
     clock: term(),
     configuration: term(),
     dumper: term(),
     executor: term(),
     executor_topic: term(),
     logger: term(),
     logger_topic: term(),
     mempool: term(),
     mempool_topic: term(),
     ordering: term(),
     pinger: term(),
     router: term(),
     storage: term(),
     storage_topic: term(),
     transport: term()
   }}
  | any()

I launch a node given a file containing a binary version of an 12-tuple with appropriate info in the following order:

router id

mempool topic id

executor topic id

dumper

storage

logger

clock

ordering

mempool

pinger

executor

storage names

qualified

order

block_storage

All engines have info on their states and id's so that checkpointing the system will keep all adresses used in the previous session. Note that I ensure that the apporpriate tables are new.

Moreover, I ensure that the mempool and block storage are in sync. In particular, I check that the order of the last block is less than that of the mempool dumped. If not, I manually remove the last block.

Check whether your transactions have had an assigned worker. If not, relaunch them directly. If blocks were out of sync with mempool, relaunch the executions as well.

Options

see launch_options/0 for the full list of optional arguments

load(name)

@spec load(String.t()) :: any() | dump()

I read the given file which I assume contains binary info and convert it to an Elixir term.

As the dumped state may have extra atoms not present in the session, I currently allow for atom creation in the loaded term.

remove_dump(file, env \\ Application.get_env(:anoma, :env))

@spec remove_dump(Path.t(), atom()) :: :ok

Removes the given dump files at the specified address and with the given configuration.

See Anoma.System.Directories for more informaiton about the path resolution and for the second atom.