Anoma.Identity.Backend.Memory (Anoma v0.11.0)

I keep the identity in memory. I may or may not persisting across reboots. This depends if the table is set to disc_only copies, or memory_copies

Types

t()

@type t() :: %Anoma.Identity.Backend.Memory{
  nonce: binary(),
  symmetric: Anoma.Crypto.Symmetric.t(),
  table: atom()
}

We determine the information needed to properly store a memory copy.

Fields

symmetric - this is a symmetric encryption/decryption key known by the user and the system. This key will be used to encrypt the public and private key, and along with the External.t/1 can index into table

table - this is the local table in which where in memory this will be stored