Anoma.Node (Anoma v0.18.0)

I act as a registry for Anoma Nodes

There are two ways to launch a Node. Either with minimal arguments or having the full specification for all the engines, including their external ID's.

All inputs should come in format of a list of atom-value 2-tuples, {:settings, map()} specifying the info on all relevant engines and storages possibly including their tables presented in a dumped format, {:name, atom()} specifying the name of the node, and {:new_storage, bool()} specifying whether the table names, if supplied, need to be set-up without any deletions.

For more info on format check start_min/1

Minimal Arguments

name - name for this process

snapshot_path : [atom() | 0]

A snapshot location for the service (used in the worker)

storage : Anoma.Node.Storage.t() - The Storage tables to use

block_storage - a location to store the blocks produced

Created Tables

storage.qualified

storage.order

block_storage

Types

configuration()

@type configuration() :: [
  name: atom(),
  use_rocks: boolean(),
  settings: node_settings(),
  testing: boolean() | nil
]

I am the configuration type for the Anoma.Node.

Ι contain information necessary for tweaking the behavior of the Anoma.Node application. See my Fields section for more information on how my behavior changes the node.

Fields

:use_rocks - determines if we wish to use rocksdb, or use mnesia in memory database

:settings - are the engine specific and anoma node configuration settings. See node_settings/0 for more details

:testing is a flag that specifies if the node being launched is for testing purposes. Currently this just affects if we take up the standard socket over the current Anoma instance or not.

engine_configuration()

@type engine_configuration() :: %{
  clock: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Clock.t()},
  configuration:
    {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Configuration.t()},
  pinger: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Pinger.t()},
  logger: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Logger.t()},
  ordering: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Ordering.t()},
  executor: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Executor.t()},
  mempool: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Mempool.t()},
  storage: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Storage.t()},
  dumper: {Anoma.Crypto.Id.Extern.t() | nil, Anoma.Node.Dumper.t()},
  storage_data: {Anoma.Node.Storage.t(), atom()},
  snapshot_path: Noun.t()
}

min_engine_configuration()

@type min_engine_configuration() :: [
  ping_time: :no_timer | non_neg_integer(),
  name: atom(),
  snapshot_path: Noun.t(),
  storage_data: Anoma.Node.Storage.t(),
  block_storage: atom(),
  configuration: Anoma.Configuration.configuration_map() | nil
]

node_settings()

@type node_settings() ::
  {:new_storage, engine_configuration()} | {:from_dump, Anoma.Dump.dump()}

I am the node settings

I can be resumed from two states, either from new/fresh storage in which I just have a engine_configuration/0 or from a previously dumped configuration. In which case I contain Anoma.Dump.dump/0.

If the mode is in :new_storage, then Ι just contain Engine specific configuration settings. However if I'm a :from_dump then Ι have a superset of that information containing mnesia table configuration information as well.

t()

@type t() :: %Anoma.Node{
  clock: Anoma.Node.Router.addr() | nil,
  configuration: Anoma.Node.Router.addr() | nil,
  dumper: Anoma.Node.Router.addr() | nil,
  executor: Anoma.Node.Router.addr() | nil,
  executor_topic: Anoma.Node.Router.addr() | nil,
  logger: Anoma.Node.Router.addr() | nil,
  logger_topic: Anoma.Node.Router.addr() | nil,
  mempool: Anoma.Node.Router.addr() | nil,
  mempool_topic: Anoma.Node.Router.addr() | nil,
  ordering: Anoma.Node.Router.addr() | nil,
  pinger: Anoma.Node.Router.addr() | nil,
  router: Anoma.Node.Router.addr(),
  storage: Anoma.Node.Router.addr() | nil,
  storage_topic: Anoma.Node.Router.addr() | nil,
  transport: Anoma.Node.Router.addr()
}

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(args)

@spec init(engine_configuration() | Anoma.Dump.dump()) :: any()

Callback implementation for GenServer.init/1.

raw_storage(node)

@spec raw_storage(t()) :: Anoma.Node.Storage.t()

I give the storage from a node.

This is useful when we want to bring storage to a new node.

start_link(args)

@spec start_link(configuration()) :: GenServer.on_start()

I assume I am fed a list with atom-value 2-tuples :name, :settings, :use_rocks. The value of :settings is a 2-tuple with the first component matching either :new_storage or :from_dump. If the former case, I need a :snapshot_path key in the settings. For the latter case, check Anoma.Dump for format descriptions of the settings.

I ensure that the storages are deleted and created. Afterwards, if the storage is truly new, I put the snapshot in the ordering. Otherwise, I simply repopulate the tables from a supplied list by the settings. If :use_rocks has value true, I use rocksdb as storage backend.

start_link_or_find_instance(args)

@spec start_link_or_find_instance(configuration()) :: GenServer.on_start()

I work just like start_link/1 except I give back the pid directly if the node is already started.

This is preferable in tests, as we wish to have them be robust upon rerunning.

start_min(args)

@spec start_min(min_engine_configuration()) :: engine_configuration()

Given minimal arguments, I create appropriate setup for the :settings argument for Node initialization.

state(nodes)