Anoma.Node.Configuration (Anoma v0.21.0)

I am the implementation of the Configuration Engine.

I remember the configuration that the application was launched with as well as call the snapshotting and snapshot-deletion functionality.

Public API

I have the following public functionality:

snapshot/1

delete_dump/1

Types

t()

@type t() :: %Anoma.Node.Configuration{
  configuration: Anoma.Configuration.configuration_map(),
  logger: Anoma.Node.Router.Addr.t() | nil
}

I am the type of the Configuration Engine.

Fields

:configuration - The configuration data stored in specified format.
               Please consult the
               `t:Anoma.Configuration.configuration_map/0`.
               Enforced: true.

:logger - The address of the Logger Engine. Enforced: false.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delete_dump(config)

@spec delete_dump(Anoma.Node.Router.addr()) :: :ok

I am the function deleting the snapshot file.

I check the dump path and check whether there is any file snapshot there. If so, I delete it, otherwise I do nothing.

handle_cast(atom, caller, config)

init(state)

@spec init(t()) :: {:ok, t()}

I am the Configuration Engine initialization function.

I receive a Configuration.t() structure and launch the engine instance with the fed-in state.

snapshot(config)

@spec snapshot(Anoma.Node.Router.addr()) :: :ok

I am the snapshot function.

I take a snapshots of the current state. The topic sends back a message to the caller saying :snapshot_done.

The path for the snapshot is taken directly from the configuration map.