Anoma.Node.Dumper (Anoma v0.21.0)

I am the Dumper Engine.

My role is to spawn asynchronous processes which listen to a specified block table announcement and await when blocks modulo my count have been executed, at which point I checkpoint the assigned node using the duping system using a file with a specified name.

Public API

I have the following public functionality:

start/1

stop/1

set_count/2

Types

t()

@type t() :: %Anoma.Node.Dumper{
  configuration: Anoma.Node.Router.Addr.t() | nil,
  count: non_neg_integer() | nil,
  logger: Anoma.Node.Router.Addr.t() | nil,
  task: Task.t() | nil
}

I am the type of the Dumper Engine.

I store the number of blocks to be passed before snapshotting, the Configuration Engine address, as well as the task responsible for the snapshotting.

Fields

:count - Field specifying how many blocks ought to pass between
       snapshots. Enforced: false.

:configuration - The Configuration Engine address using which the
               snapshot path is found.

:task - The field containing the task carrying out the
      snapshotting. Enforced: false.

:logger - The address of the Logger Engine. Enforced: false.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_cast(arg1, from, state)

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  count: non_neg_integer() | nil,
  configuration: Anoma.Node.Router.Addr.t(),
  logger: Anoma.Node.Router.Addr.t() | nil
) :: {:ok, t()}

I am the initialization function for the Dumper Engine.

Pattern-Matching Variations

init(%Dumper{}) - I initialize the Engine with given state.

init(opts) - I expect a keylist with the :count, :configuration,
           and `:logger` keys available. I then start the engine
           instance with appropriate state.

set_count(server, count)

@spec set_count(Anoma.Node.Router.Addr.t(), non_neg_integer() | nil) :: :ok

I set the count of a Dumper Engine instance.

I first shutdown the current task, start a new task with a changed count parameter and return and register both the count and the task changes.

If count is a non-positive integer, I do nothing. If it is nil, I set both the count and the task at hand to nil.

start(server)

@spec start(Anoma.Node.Router.Addr.t()) :: :ok

I am the Dumper start function.

I start a dumping loop based on the count supplied. If an old task was present before the start, I shut it down. If the count is nil I return a state with nil count and nil task.

stop(server)

@spec stop(Anoma.Node.Router.Addr.t()) :: :ok

I am the Dumper stop function.

I shutdown the dumper task registered in the struct. Depending on the shutdown message, I provide separate logs and then set the task field to nil.