Anoma.Node.Examples.EIntentPool (Anoma v0.28.0)

I contain several examples on how to run the intent pool.

Functions

add_intent(enode \\ ENode.start_node())

@spec add_intent(Anoma.Node.Examples.ENode.t()) :: Anoma.Node.Examples.ENode.t()

I check that when an intent is added to the pool, is is present in the mapset.

add_intent_transaction_nullifier(enode \\ ENode.start_node())

@spec add_intent_transaction_nullifier(Anoma.Node.Examples.ENode.t()) ::
  Anoma.Node.Examples.ENode.t()

I submit a transaction containing a nullifier to an ephemeral resource used in a trivial swap transaction.

list_intents(enode \\ ENode.start_node())

@spec list_intents(Anoma.Node.Examples.ENode.t()) :: Anoma.Node.Examples.ENode.t()

I check that the intent pool returns an empty map of intents when its started.

remove_intent(enode \\ ENode.start_node())

@spec remove_intent(Anoma.Node.Examples.ENode.t()) :: Anoma.Node.Examples.ENode.t()

I add and remove an intent from the pool and ensure that it's actually gone.

remove_intents_with_nulllified_resources(enode \\ ENode.start_node())

@spec remove_intents_with_nulllified_resources(Anoma.Node.Examples.ENode.t()) ::
  Anoma.Node.Examples.ENode.t()

I use add_intent_transaction_nullifier and then execute the trivial swap.

The nullifier of the ephemeral resource used gets trasmitted to the intent pool, hence removing the specified intent from the pool.