Anoma.Node.Executor (Anoma v0.11.0)

I Manage the Pub Sub behavior

If new intents come in, I send it to all my subscribers.

Further I have the job of spawning Executor tasks when new transactions come in

Currently I only communicate:

When tasks are completed

API

My public facing API is

new_transaction/3

new_transaction/4

fire_new_transaction/3

fire_new_transaction/4

snapshot/1

state/1

subscribe/2

Types

t()

@type t() :: %Anoma.Node.Executor{
  ambiant_env: Nock.t() | nil,
  logger: Anoma.Node.Router.Addr.t() | nil,
  task_completion_topic: Anoma.Node.Router.Addr.t() | nil,
  tasks: [Task.t()]
}

Functions

fire_new_transaction(executor, order, gate)

@spec fire_new_transaction(
  Anoma.Node.Router.Addr.t(),
  Noun.t(),
  Anoma.Transaction.execution()
) :: pid()

Acts like new_transaction/3, but the caller does not care about the response. Instead the response is handled by the pool.

The user gets back a process so it may keep track of sending messages to this task or terminating the task

fire_new_transaction(executor, order, gate, env)

@spec fire_new_transaction(Anoma.Node.Router.Addr.t(), Noun.t(), Noun.t(), Nock.t()) ::
  pid()

handle_call(arg1, from, state)

handle_cast(atom, agent)

handle_info(process, state)

init(state)

kill_transactions(communicator, trans)

new_transaction(executor, order, gate)

@spec new_transaction(Anoma.Node.Router.Addr.t(), Noun.t(), Noun.t()) :: Task.t()

Spawns a new transaction, the returned task's owner is the caller.

Inputs

executor - the genserver to send the message to

order - the unique ID of the transaction

gate - a Nock function that we will spawn a task for

Output

a task that the caller owns

Example

we assume a running Worker, see executor_test.exs for a full example

> id = System.unique_integer([:positive])
1931
> zero = [[1, 777 | 0], 0, Nock.stdlib_core()]
> task = Executor.new_transaction(executor, id, {:kv, zero})
%Task{
  mfa: {Anoma.Node.Executor.Worker, :run, 3},
  owner: #PID<0.264.0>,
  pid: #PID<0.753.0>,
  ref: #Reference<0.0.33795.904691850.4087414796.24378>
}

new_transaction(executor, order, gate, env)

@spec new_transaction(Anoma.Node.Router.Addr.t(), Noun.t(), Noun.t(), Nock.t()) ::
  Task.t()

snapshot(executor)

Returns the snapshot path, the Nock code is using

Example

iex> alias Anoma.Node.{Executor, Router} iex> {:ok, router} = Router.start iex> snap = %Nock{snapshot_path: [:a | 0], ordering: nil} iex> args = {snap, Router.new_topic(router), nil} iex> {:ok, addr} = Router.start_engine(router, Executor, args) iex> Executor.snapshot(addr) :a

state(executor)

Returns the current state of the executor