Anoma.Node.Executor.Worker (Anoma v0.19.0)

I am the Worker Engine.

My instance gets launched by the Executor and is connected to a unique transaction.

I am responsible for the main work done to run a succesful transaction lifecycle. This includes processing of transactions, calling for their ordering via the Ordering Engine, creation and revision of commitment trees, nullifier key-checking, as well as the storage of relevant data for transaction completion before block-execution.

Public API

I provide the following public functionality:

rm_nullifier_check/2

Types

backend()

@type backend() :: :kv | :rm | :cairo | :ro

from_noun(trans)

@type from_noun(trans) :: (Noun.t() -> trans)

t()

@type t() :: %Anoma.Node.Executor.Worker{
  completion_topic: Anoma.Node.Router.Addr.t() | nil,
  env: Nock.t() | nil,
  id: non_neg_integer() | nil,
  reply_to: Anoma.Node.Router.addr() | nil,
  tx: transaction() | nil
}

I am the type of a Worker Engine instance.

I contain all the info for appropriate transaction processing.

Fields

:id - The ID of the transaction fed in.

:tx - The transaction code.

:env - The environment for the transaction to be evaluated in. E.g.
     contains the Ordering engine address. See `Nock.t()`

:completion_topc - The address of the topic connected to the
                 relevant Executor Engine for broadcasting.

:reply_to - The address that the computed value (for read-only
          transaction) is sent to.

transaction()

@type transaction() :: {backend(), Noun.t() | binary()}

verify_fun(trans)

@type verify_fun(trans) :: (trans -> boolean())

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(arg)

@spec init(
  {non_neg_integer(), transaction(), Nock.t(), Anoma.Node.Router.Addr.t(),
   Anoma.Node.Router.addr() | nil}
) :: {:ok, t()}

I am the Worker initialization function.

I send myself a :run message which launches my core functionality and return the appropriate state.

Pattern-Matching Variations

init({id, tx, env, completion_topic, reply_to}) - I receive a tuple with all
                                                the specified info to launch
                                                a Worker instance.

rm_nullifier_check(storage, nullifiers)

@spec rm_nullifier_check(Anoma.Node.Router.addr(), [binary()]) :: bool()

I perform the nullifier check for a resource machine transaction.

Given a storage and a list of nullifiers I check their placing in storage.