Anoma.Node.Logger (Anoma v0.19.0)

I am the Logger Engine, an implementation of the Local Logging Engine.

I have a storage field which accepts the storage used for the logging tied to a specific node, as well as the address of the clock used for timestamping.

I store and get the logging info using the keyspace method.

Public API

I provide the following public functionality:

Adding Logs

add/3

Getting Logs

get/1

get/2

Types

t()

@type t() :: %Anoma.Node.Logger{
  clock: Anoma.Node.Router.Addr.t() | nil,
  storage: Anoma.Node.Router.Addr.t() | nil,
  topic: Anoma.Node.Router.Addr.t() | nil
}

I am the type of the Logging Engine.

I have the minimal fields required to store timestamped messages in the specified storage.

Fields

:storage - The Storage Engine address used for storing.

:clock - The Clock Engine address used for timestamping.

:topic - The topic address for broadcasting.

Functions

add(logger, atom, msg)

@spec add(Anoma.Node.Router.addr() | nil, atom(), String.t()) :: :ok | nil

I am the Logger add function, the implementation of LocalLoggingAppend of the Local Logging Engine of the Anoma Specification.

I receive three arguments: an address of the logger, an atom specifying the urgency of the logging message and the logging message itself.

If the logger address is indeed an address, I put the info inside the specified storage using Storage.put/3.

The information is stored using a keyspace method. The message is hence attached to a list of 4 elements:

ID of the logger calling for storage

ID of the engine whose info we log or a PID of the worker process

Timestamp relative to the clock engine used

The urgency atom

I then use the atom to call the Elixir logger in order to inform the user of any specific major logging event such as, e.g. failing workers.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get(logger)

@spec get(Anoma.Node.Router.addr() | nil) ::
  [{[Anoma.Crypto.Id.Extern.t() | integer() | atom()], String.t()}] | nil

I am the Logger get function of one argument.

Given a non-nil logger address, I get the entire keyspace related to the storage attached to the aforementioned logger ID using Storage.get_keyspace/2

I return a list of 2-tuples. Its left element will be a list whose head is the ID of the specified logger, followed by the engine ID (or worker PID), the timestamp, and an atom specifying urgency of the logging message. Its right element will be the message.

get(logger, engine)

@spec get(Anoma.Node.Router.addr() | nil, Anoma.Node.Router.addr() | pid()) ::
  [{[Anoma.Crypto.Id.Extern.t() | integer() | atom() | pid()], String.t()}]
  | nil

I am the Logger get function of two arguments.

Given a non-nil logger address and an engine address, I get the keyspace which conatins all the info stored by the logger about the supplied engine using Storage.get_keyspace/2

I return a list of 2-tuples. Its left element will be a list whose head is the ID of the specified logger, followed by the engine ID (or worker PID), the timestamp, and an atom specifying urgency of the logging message. Its right element will be the message.

handle_cast(arg, addr, state)

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  storage: Anoma.Node.Storage.t(),
  clock: Anoma.Node.Router.addr(),
  topic: Anoma.Node.Router.addr()
) :: {:ok, t()}

I am the Logger Engine initialization function.

Pattern-Matching Variations

init(Logger.t()) - I initialize the Engine with the given state.

init(args) - I expect a keylist with :storage, :clock, and
           `:topic` keys and launch the Engine with the appropriate
           state.