Anoma.Node.Mempool (Anoma v0.20.0)

I am the Mempool Engine.

I provide the interface for transaction management including communication with the Executor engine to create transaction candidates with given code, executing said transactions, and creating appropriate blocks.

Public API

I provide the following public functionality:

execute/1

soft_reset/1

hard_reset/1

tx/2

Types

t()

@type t() :: %Anoma.Node.Mempool{
  block_storage: atom(),
  executor: Anoma.Node.Router.Addr.t() | nil,
  key: {Anoma.Serializer.public_key(), Anoma.Serializer.private_key()},
  logger: Anoma.Node.Router.Addr.t() | nil,
  ordering: Anoma.Node.Router.Addr.t() | nil,
  round: non_neg_integer(),
  topic: Anoma.Node.Router.Addr.t() | nil,
  transactions: transactions()
}

I am the type of the Mempool Engine.

My fields store information necessary to succesfully call for transaction candidate creation, execution, and ordering requests.

Fields

:ordering - The Ordering Engine Address.

:executor - The Executor Engine Address.

:block_storage - The name of the storage for created blocks.
               Default: `Anoma.Block`

:transactions - List of transaction candidates. Currently these do
              not have index info. That is handled by the
              Ordering engine.
              Default: []

:round - The block round information.
       Default: 0

:topic - The topic address for broadcasting.

:key - The key used for block creation.
     Default: `:crypto.generate_key(:rsa, {1024, 65537}`

:logger - The address of the Logger Engine.
        Enforced: false

transactions()

@type transactions() :: [Anoma.Transaction.t()]

I am a list of transaction candidates.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

choose_and_execute_ordering(state, transactions)

@spec choose_and_execute_ordering(t(), [Anoma.Transaction.t()]) :: :ok

I order the transactions based on the current ordering specified by the Ordering engine, place a new orderig, and for every ordered transaction, send to their appropriate Worker a :write_ready message.

Return :ok.

execute(server)

@spec execute(Anoma.Node.Router.Addr.t()) :: :ok

I am the execution function.

I execute the block containing top 100 transaction candidates availiable, ordering them using the specified Ordering Engine.

handle_cast(arg1, from, state)

handle_execute(state)

@spec handle_execute(t()) :: {non_neg_integer(), t()}

I handle Mempool execution.

I take the top 100 transaction candidates, ask for an ordering on them, producing and storing an encoded block.

I return a tuple with first entry specifying how many transactions got into the execution and a new appropriately changed state.

handle_tx(tx_code, state, reply_to)

@spec handle_tx(Anoma.Transaction.execution(), t(), Anoma.Node.Router.addr() | nil) ::
  Anoma.Transaction.t()

I handle the transaction candidate creation.

Given transaction code and the Mempool state, I ask the Executor to fire a new transaction with a randomly generated transaction ID.

I reply with a newly created transaction and specified Worker.

hard_reset(server)

@spec hard_reset(Anoma.Node.Router.Addr.t()) :: :ok

I am the hard reset function.

Similar to soft_reset/1, I kill transactions, but also call for an Executor snapshot, delete the block table and hard reset the Ordering Engine. The block table is then re-launched with the same rocks flag as before.

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  ordering: Anoma.Node.Router.Addr.t(),
  executor: Anoma.Node.Router.Addr.t(),
  block_storage: atom(),
  round: non_neg_integer(),
  topic: Anoma.Node.Router.Addr.t(),
  key: {Anoma.Serializer.public_key(), Anoma.Serializer.private_key()},
  logger: Anoma.Node.Router.Addr.t()
) :: {:ok, t()}

I am the initialization function for a Mempool Engine instance.

Pattern-Macthing Variations

init(%Mempool{}) - I initialize the Engine with the given state.

init(args) - I expect a keylist with all keys in the structure fields
           listed necessary for the Mempool Engine to start. I then
           create a table with the given `:block_storage` argument
           and initialize an instance with given parameters.

kill_transactions(state)

@spec kill_transactions(t()) :: :ok

I kill transactions listed in the fed-in state.

Given a state with a list of transaction candidates, I call for the Executor engine to kill each one in the list.

order(executing, next_order)

@spec order([Anoma.Transaction.t()], non_neg_integer()) ::
  Anoma.Node.Ordering.ordered_transactions()

I order a list of transactions based on a given order.

Given a list of transaction candidates and an order number, I shuffle the transactions, and index them starting with the given order number.

order_format(transaction, order)

@spec order_format(Anoma.Transaction.t(), non_neg_integer()) :: Anoma.Transaction.t()

I order a transaction structure by giving it an index.

Given a transaction candidate and an order, I put the index specified by the input as its :index field value.

persistent_transaction(trans)

@spec persistent_transaction(Anoma.Transaction.t()) ::
  {Noun.t(), Anoma.Transaction.execution()}

I give core information on a transaction candidate.

Given a transaction candidate, I return a tuple. The first element is the id of the transaction. The second element is the transaction code.

produce_block(state, trans)

@spec produce_block(t(), [Anoma.Transaction.t()]) :: Anoma.Block.t()

I produce a block with the key and round from the given state and eturn it.

reset_blocks(state)

@spec reset_blocks(t()) :: {:aborted, any()} | {:atomic, :ok}

I provide block-resetting functionality.

I delete the mnesia table for block storage provided in the given state.

I then create the table with the same name and rocks flag.

reset_state(state)

@spec reset_state(t()) :: t()

I return a state with no transactions and 0 round count.

save_block(state, block)

@spec save_block(t(), Anoma.Block.t()) :: {:atomic, :ok} | {:aborted, any()}

I save a provided block.

Given a Mempool Engine state and a block, I encode the block and write it at the specified mnesia table.

soft_reset(server)

@spec soft_reset(Anoma.Node.Router.Addr.t()) :: :ok

I am the soft reset function.

I empty the transaction and round fields of the appropriate Mempool Engine, asking the linked Executor to kill relevant transaction Workers.

tx(server, tx_code, reply_to \\ nil)

@spec tx(
  Anoma.Node.Router.Addr.t(),
  Anoma.Transaction.execution(),
  Anoma.Node.Router.addr() | nil
) ::
  :ok

I am the transaction candidate creating API.

Given a Mempool Engine address and a transaction candidate code, I ask the Executor Engine to fire a new transaction with a random ID created in the process. The value computed by the transactions (for read-only transactions) will be sent to reply_to address. I return the new state, adding the transaction.