Anoma.Node.Ordering (Anoma v0.18.0)

I am the Ordering Engine.

My main functionality is to calculate transaction ordering and keep track of said odering relative to the transaction id's.

Public API

I provide the following public functionality:

Ordering

true_order/2

new_order/2

Reset

reset/1

hard_reset/2

Blocking

caller_blocking_read_id/2

Other

handle_new_order/2

Types

key()

@type key() :: any()

I am a type representing a key in a key-value map.

ordered_transactions()

@type ordered_transactions() :: [Anoma.Transaction.t()]

I am a list of ordered transactions.

t()

@type t() :: %Anoma.Node.Ordering{
  hash_to_order: %{required(key()) => non_neg_integer()},
  logger: Anoma.Node.Router.Addr.t() | nil,
  next_order: non_neg_integer(),
  storage: Anoma.Node.Router.Addr.t() | nil
}

I am the type of the Ordering Engine.

I have basic fields relating to upcoming ordering for the incoming transactions as well as mapping of transaction ID's to their orders.

Fileds

:storage - The address of the storage engine

:next_order - The integer referencing the order to be given to the
            next incoming transaction.
            Default: 1

:hash_to_order - A map of transaction ID's and their orders.
               Default: %{}

:logger - The Logger Engine address.

Functions

caller_blocking_read_id(ordering, list)

@spec caller_blocking_read_id(Anoma.Node.Router.Addr.t(), Noun.t()) ::
  :error | {:ok, any()}

I call the blocking read functionality on a Worker request.

I am called by the Worker, who sends me an Ordering Engine address and a list with head a transaction ID and a subkey for the storage key.

I get the orders of the transactions via checking their ID keys in the Ordering Engine. I then return the list by interchanging the transaction ID with its order and call Storage.blocking_read/2 with the appropriate arguments.

Pattern-Matching Variations

caller_blocking_read_id(ordering, [id | subkey]) - I match on the

                                                 given list to
                                                 check transaction
                                                 ID.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_cast(arg1, from, state)

handle_new_order(ordered_transactions, state)

@spec handle_new_order(ordered_transactions(), t()) ::
  {non_neg_integer(), %{required(key()) => non_neg_integer()}}

I handle the new ordered transactions comming in to the Ordering Engine.

I send read_ready messages regarding the transaction orders, then add the length of the incoming transaction list to the :next_order value, finally updating the :hash_to_order map with new key-values.

I return a tuple where the first argument is the new proposed next order while the second argument is the new map of ID's to orders.

hard_reset(ordering, initial_snapshot)

@spec hard_reset(Anoma.Node.Router.Addr.t(), atom()) :: :ok

I hard reset the Ordering Engine.

Similarly to reset/1 I get rid of all hot state fields in the Ordering Engine. Moreover, I ask the Storage engine to ensure that the tables are re-launched and put a new snapshot specified by the second argument.

init(state)

@spec init(t()) :: {:ok, t()}

@spec init([
  {:storage, Anoma.Node.Router.Addr.t()}
  | {:logger, Anoma.Node.Router.Addr.t()}
  | any()
]) ::
  {:ok, t()}

I am the initialization function for a Ordering Engine instance.

Pattern-Macthing Variations

init(%Ordering{}) - I initialize the Engine with the given state.

init(args) - I expect a keylist with the :logger and :storage
           keys. I then ask the Storage engine given by the keyword
           to setup and return the appropriate initialization
           state.

new_order(ordering, ordered)

@spec new_order(Anoma.Node.Router.Addr.t(), ordered_transactions()) :: :ok

I am the function dealing with new ordered transactions.

Given a list of ordered transactions, I update the Ordering Engine state appropritaely by changing the :next_order field and filling in the :hash_to_order field with new key-value pairs pairing the new ID's with their asigned orderings.

reset(ordering)

@spec reset(Anoma.Node.Router.Addr.t()) :: :ok

I am the Ordering Engine reset function.

I get rid of all hot state fields of the Ordering Engine leaving only the Storage and Logger address intact.

true_order(ordering, id)

@spec true_order(Anoma.Node.Router.Addr.t(), any()) :: non_neg_integer() | nil

I am the true order function.

Given a transaction ID, I get the order associated with the transaction. I look at the map stored in the state and then check the value of said ID.