Anoma.Node.Pinger (Anoma v0.18.0)

I am the Pinger Engine.

I provide periodic block execution based on submitted mempool address and time by calling the execution API in the mempool engine.

Public APIs

I have the following public functionality:

start/1

pinger/1

set_timer/2

Types

t()

@type t() :: %Anoma.Node.Pinger{
  logger: Anoma.Node.Router.Addr.t() | nil,
  mempool: Anoma.Node.Router.Addr.t() | nil,
  time: non_neg_integer() | atom()
}

I am the type of the Pinger Engine instance.

I store minimal info required to ask the mempool to execute, namely the mempool address and the time specified by the user.

:mempool - The Mempool Engine address which is called to execute.

:time - The time that should be elapsed between the calls to
      execute or an atom saying that no timer should be set.
      Default: `:no_timer`

:logger - The address of the Logger Engine.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_cast(arg1, from, state)

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  mempool: Anoma.Node.Router.Addr.t(),
  time: non_neg_integer(),
  logger: Anoma.Node.Router.Addr.t()
) :: {:ok, t()}

I am the initialiation function for the Pinger Engine.

Patern-Matching Variations

init(%Pinger{}) - I initialize the Engine with the given state.

init(args) - I expect a keylist with the :mempool, :time and
           `:logger` keywords and then start a Pinger Engine
           instance with appropriate state.

pinger(time)

@spec pinger(:no_timer | non_neg_integer()) :: :ok | reference()

I am the pinger function.

I receive an argument which is either an integer or :no_timer. If it is an integer, send to self an :execute message after specified ammount of time. Otherwise, simply reply :ok

set_timer(server, time)

@spec set_timer(Anoma.Node.Router.Addr.t(), non_neg_integer() | :no_timer) :: :ok

I set the timer field for a Pinger Engine instance.

Given a server S and time T I change the timer set for the struct connected to S setting it to T. Set T to :no_timer to stop the pinger.

start(server)

@spec start(Anoma.Node.Router.Addr.t()) :: :ok

I start the Pinger Engine's execution-calling functionaluty.

Given a pinger address, I start up the pinger by calling the pinger/1 function feeding it the time associated to the address.

Note that if the timer specified has value :no_timer, the pinger will not practically start.