Anoma.Node.Router (Anoma v0.12.0)

I have a few use cases and APIs throughout an Anoma application:

I provide central routing infrastructure for inter-node and intra-node communication

I provide service's for topic creation and Anoma.Node.Router.Engine creation

I provide a behavior interface similar to GenServer for both Engine messaging and Topic messaging

Router Topology

A good view of my topology and how various Anoma.Node.Router.Engine and topic's relate to me can be seen by the following diagram.

graph TB
  S(Router Supervisor)
  R(Router)
  C(Engine #1) ~~~ B(Engine #2)
  T1(Topic #1) ~~~ T2(Topic #2)
  S ==>R & B & C
  S -.->T1 & T2
  R <-->|router| B & C & T1 & T2

Anoma.Node.Router.Engine's that are spawned via start_engine/3 are added to the supervisor, and messages sent via call/3 or cast/2 may be routed through myself.

Topics are handled somewhat specially. Since they serve a much more limited role, I handle and manage them personally and the dashed lines in the diagram are not real. However for building a conceptual model, we can view Topics spawned by new_topic/1 as spawning an Engine, and messages sent via call/3 or cast/2 as being routed through myself.

Server Terminology

For the sake of this document, the word Server will refer to both Engines and Topics.

Router PIDs

The IDs of the various topics and engines, are not regular Erlang PID's. Instead we use our own address schema laid out in Anoma.Node.Router.Addr. Further information can also be found in our specs (link to specs when they are online).

Router API

I offer an API for anyone wishing to call me. These functions are:

start/0

start/1

new_topic/1

start_engine/3

start_engine/4

subscribe_topic/2

subscribe_topic/3

Server APIs

When writing server code, My module acts as a GenServer behavior, in that one should be using these functions when writing code to talk to the Server:

call/2

call/3

cast/2

The API I offer is very reminiscent of GenServer's API, however please read the examples section to get a sense on how the server code differs.

Client/Server APIs

Much like GenServer, user's typically don't call my call/3 or cast/2 directly. Instead the user wraps the calls in new functions representing the public API of the server. These wrappers are called the client API.

Examples

To get a good feel for writing idiomatic code using me, and how my API is differs from GenServer, we will look at an implementation of a stack and see how it differs from GenServer's stack example

defmodule Stack do
  alias Anoma.Node.Router

  use Router.Engine

  @spec init(String.t()) :: {:ok, list(String.t())}
  def init(elements) do
    initial_state = String.split(elements, ",", trim: true)
    {:ok, initial_state}
  end

  @spec push(Router.addr(), String.t()) :: :ok
  def push(pid, element) do
    Router.cast(pid, {:push, element})
  end

  @spec pop(Addr.t()) :: String.t()
  def pop(pid) do
    Router.call(pid, :pop)
  end

  def handle_call(:pop, _from, state) do
    [to_caller | new_state] = state
    {:reply, to_caller, new_state}
  end

  def handle_cast({:push, element}, _from, state) do
    new_state = [element | state]
    {:noreply, new_state}
  end
end

And we can use this Engine like the following

iex> {:ok, router} = Router.start()
iex> {:ok, stack} = Router.start_engine(router, Stack, "hello")
iex> Stack.pop(stack)
"hello"
iex> Stack.push(stack, "hi")
:ok
iex> Stack.pop(stack)
"hi"

Overall having both server and client functionality that is clearly labeled, leads to good user UI's.

Difference in Design from GenServer

The Stack example shows how my API differs from Genserver:

There is currently no GenServer.start_link/3 for the Router

Even handle_cast's take a from parameter

We use Anoma.Node.Router.Engine and not GenServer

Further call/2 by default has an :infinite timeout

Summarizing the interactions between Server and Clients

For basic Server's we can summarize interactions between all parties as follows.

  sequenceDiagram
    participant C as Client (Process/Engine)
    participant R as Router (Engine)
    participant S as Server (Engine)
    participant M as Module (Code)

    note right of C: Typically started by an init process
    C->>+R: Router.start_engine(router, module, arg, options)
    R->>+S: GenServer.start_link(router, module, arg, options)
    S-->>+M: init(arg)
    M-->>-S: {:ok, state} | :ignore | {:error, reason}
    S->>-R: {:ok, addr} | :ignore | {:error, reason}
    R->>-C: {:ok, addr} | :ignore | {:error, reason}

    note right of C: call is synchronous
    C->>+R: Router.call(addr, message)
    note right of R: For known engines this is optimized away
    R->>+S: GenSever.call(pid, {self, message})
    S-->>+M: handle_call(message, from, state)
    M-->>-S: {:reply, reply, state} | {:stop, reason, reply, state}
    S->>-C: reply

    note right of C: cast is asynchronous
    C-)R: Router.cast(addr, message)
    note right of R: For known engines this is optimized away
    R-)S: GenServer.cast(pid, message)
    S-->>+M: handle_cast(message, state)
    M-->>-S: {:noreply, state} | {:stop, reason, state}

    note right of C: send is asynchronous. the PAID must be known
    C-)S: Kernel.send(pid, message)
    S-->>+M: handle_info(message, state)
    M-->>-S: {:noreply, state} | {:stop, reason, state}

What this diagram doesn't show is how Topic messaging works, and for that, this diagram may be of assistance.

  sequenceDiagram
    participant C as Client (Process/Engine)
    participant S2 as Server2 (Engine)
    participant R as Router (Engine)
    participant T as Topic (Topic)
    participant S as Server (Engine)

    note right of R: For simplicity we do not display a client
    S2-)+R: GenServer.subscribe(router, topic)
    R-->>+T: GenSever.call(addr, {self, message})
    T-->>-R: :ok | {:error, :no_such_topic}
    R-)-S2: :ok | {:error, :no_such_topic}

    note right of C: cast is asynchronous
    C-)R: Router.cast(addr, message)
    R-)+S: GenServer.cast(pid, message)
    note right of S: Imagine this call sends an update to a topic
    S-)-R: Router.cast(topic_addr, msg)
    R-->>T: Genserver.cast(pid, msg)
    T-->>R: Router.cast(addr_S2, msg)
    R-)S2: Genserver.cast(pid, msg)

The dotted lines in the diagram are theoretically how Tasks could be implemented, however this is implementation dependent, and does indeed differ from the diagram.

Types

addr()

@type addr() :: Anoma.Node.Router.Addr.t()

t()

@type t() :: %Anoma.Node.Router{
  addr: addr() | nil,
  id: Anoma.Crypto.Id.Extern.t() | nil,
  local_engine_subs: %{required(addr()) => Anoma.Crypto.Id.Extern.t()},
  local_engines:
    %{required(Anoma.Crypto.Id.Extern.t()) => Anoma.Crypto.Id.t()} | nil,
  msg_queue: map(),
  supervisor: atom() | nil,
  topic_table: %{
    required(Anoma.Crypto.Id.Extern.t()) => MapSet.t(Anoma.Node.Router.Addr.t())
  }
}

Functions

call(addr, msg)

@spec call(Anoma.Node.Router.Addr.t(), term()) :: term()

See call/3 for documentation.

call(addr, msg, timeout)

@spec call(Anoma.Node.Router.Addr.t(), term(), :erlang.timeout()) :: term()

Makes a synchronous call to the Server and waits for a reply.

Call has a few interesting cases we can consider

Calling a local Engine

Calling a non local Engine

We can not call/3 a Topic, and thus those instances are not handled.

Currently we do not handle the non local Engine case.

For the local Engine, then their Engine.handle_call/3 will be called on the Engine to handle the request.

Timeouts

By default the timeout is infinite for local calls. For non local calls the timeout is not yet defined and will be iterated upon in future versions

cast(addr, msg)

@spec cast(Anoma.Node.Router.Addr.t(), term()) :: :ok

Makes a synchronous call to the Server and waits for a reply.

Call has a few interesting cases we can consider

Casting to a local Engine

Calling to a non local Engine

Casting to a Topic

For the local Engine, then their Engine.handle_cast/3 will be called on the Engine to handle the request.

Calling a non local Engine isn't handled yet.

For Topics any cast/2 sent, then triggers a series of cast/2 to all subscribed Process's

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_self_call(arg1, addr, s)

init(id)

@spec init(Anoma.Crypto.Id.t()) :: {:ok, t()}

Callback implementation for GenServer.init/1.

new_topic(router)

@spec new_topic(Anoma.Node.Router.Addr.t()) ::
  {:ok, Anoma.Node.Router.Addr.t()} | {:error, :already_exists}

Creates a new topic. Takes the address of the router.

new_topic(router, id)

@spec new_topic(Anoma.Node.Router.Addr.t(), Anoma.Crypto.Id.t()) ::
  {:ok, Anoma.Node.Router.Addr.t()} | {:error, :already_exists}

process_name(module, id)

@spec process_name(atom(), Anoma.Crypto.Id.Extern.t()) :: atom()

self_addr(addr)

@spec self_addr(Anoma.Node.Router.Addr.t()) :: Anoma.Node.Router.Addr.t()

start()

@spec start() :: :ignore | {:error, any()} | {:ok, Anoma.Node.Router.Addr.t()}

Starts a new router, with a random Anoma.Crypto.Id.

start(id)

@spec start(Anoma.Crypto.Id.t()) ::
  :ignore
  | {:error, {:already_started, pid()} | :max_children | term()}
  | {:ok, Anoma.Node.Router.Addr.t()}

Starts a new router with a given cryptographic ID.

start_engine(router, module, arg)

@spec start_engine(Anoma.Node.Router.Addr.t(), atom(), any()) ::
  {:ok, Anoma.Node.Router.Addr.t()} | :ignore | {:error, any()} | any()

start_engine(router, module, id, arg)

@spec start_engine(Anoma.Node.Router.Addr.t(), atom(), Anoma.Crypto.Id.t(), term()) ::
  {:ok, Anoma.Node.Router.Addr.t()}
  | :ignore
  | {:error, {:already_started, pid()} | :max_children | term()}
  | any()

Starts a new Engine

start_link(id)

@spec start_link(Anoma.Crypto.Id.t()) :: GenServer.on_start()

stop(router)