Anoma.Node.Router.Engine behaviour (Anoma v0.19.0)

I am the general Engine module.

I possess info on how all Engines are set up and process messages through the Router. Moreover, I provide access to Engine-wide API.

Note other that the coordination info, such as a Router address and module name of an engine instance, I provide an actual storage of an Engine state and all state transformaations correspond to the changes one can observe in my code.

Public API

I provide the following public functionality:

get_state/1

get_router/1

child_spec/1

terminate/2

Types

from()

@type from() :: Anoma.Node.Router.addr()

I am the type designating the Address of where the message got sent from.

t()

@type t() :: %Anoma.Node.Router.Engine{
  module: module() | nil,
  module_state: term() | nil,
  router_addr: Anoma.Node.Router.addr() | nil
}

I am the type of an Engine as seen from the Router on the Anoma networking layer.

Fields

:router_addr - The address of the Router that the Engine is
             connected to.

:module - The module name of the Engine.

:module_state - The current state of an Engine instance.

Callbacks

handle_call(request, from, state)

@callback handle_call(request :: term(), from(), state :: term()) ::
  {:reply, reply, new_state}
  | {:reply, reply, new_state,
     timeout() | :hibernate | {:continue, continue_arg :: term()}}
  | {:noreply, new_state}
  | {:noreply, new_state,
     timeout() | :hibernate | {:continue, continue_arg :: term()}}
  | {:stop, reason, reply, new_state}
  | {:stop, reason, new_state}
when reply: term(), new_state: term(), reason: term()

handle_cast(request, from, state)

@callback handle_cast(request :: term(), from(), state :: term()) ::
  {:noreply, new_state}
  | {:noreply, new_state,
     timeout() | :hibernate | {:continue, continue_arg :: term()}}
  | {:stop, reason :: term(), new_state}
when new_state: term()

Functions

child_spec(init_arg)

@spec child_spec({Anoma.Node.Router.Addr.t(), module(), any(), any()}) :: %{
  id: Anoma.Node.Router.Engine,
  start: {Anoma.Node.Router.Engine, :start_link, [any()]}
}

I am an Engine-wide child specification function.

Given a tuple with router, module, ID, argument information, I call the child_spec/1 function of the specified module and then merge with it base information where the start info gets provided from the supplied arguments.

get_router(addr)

@spec get_router(Anoma.Node.Router.addr()) :: Anoma.Node.Router.addr() | any()

I am an Engine-wide function for getting the router address associated with the given Engine.

I provide it by simply dumping the :router field of the asssociated Engine server.

get_state(addr)

@spec get_state(Anoma.Node.Router.addr()) :: struct() | any()

I am an Engine-wide function for getting the current Engine state.

Given an address of an Engine, I simply dump the state associated with it stored in the Engine structure associated with the given server name.

handle_continue(arg, state)

@spec handle_continue(term(), t()) :: {:noreply, t()} | {:stop, term(), t()}

Callback implementation for GenServer.handle_continue/2.

init(arg)

@spec init(
  {Anoma.Node.Router.addr(), atom(),
   Anoma.Crypto.Id.t() | Anoma.Crypto.Id.Extern.t(), term(), atom()}
) ::
  :ignore
  | {:ok, t()}
  | {:ok, any(),
     :hibernate | :infinity | non_neg_integer() | {:continue, any()}}
  | {:stop, any()}

I am an Engine instance initialization function. I setup the initial state of an Engine structure based on the state of an actual Engine instance.

I first cast a message to initialize a local engine to the connected eoutwe, then store the engine ID, server, and router information in the Process dictionary.

Afterwards, I call the actual (module) Engine instance initialization function, and put it as the initial state alongside the router and module information directly provided on startup.

start_link(arg)

@spec start_link(
  {Anoma.Node.Router.addr(), atom(),
   Anoma.Crypto.Id.t() | Anoma.Crypto.Id.Extern.t(), term()}
) :: :ignore | {:error, any()} | {:ok, pid()}

I am the starting function for an Engine instance.

I create an appropriate server name using Router.process_name with the supplied module name and Anoma ID, launching a GenServer with said name afterwards.

terminate(reason, state)

@spec terminate(reason, t()) :: {:stop, reason, t()} when reason: term()

I am an engine termination function.

I provide the main functionality to shut down an engine.

I first use the GenSercer functionality to route a message to clean up the local engine info stored inside the specified Router and then stop the Engine process with a given reason.