Anoma.Node.Storage (Anoma v0.18.0)

I am the Anoma Storage Engine.

The Anoma database usage is separated into two parts:

The Ordering table provides info on the order of appropriate keys. That is, it has info about how many times the key has been assigned to some value.

The Qualified table provides a representation of all stored key-values in a particular format. Each entry has inside it

The table name attached

The order of the key given by the Ordering

The namespace idenifying the Storage Engine

The key itself as provided by the user.

As a Storage Engine, I provide the appropriate functionality to coordinate the Ordering and Qualified tables of a launched Anoma Node.

That is, I help with writing things respecting the ordering of events while providing appropriate reading capabilities of the tables used, so that the base read events read only the latest value of a given key.

Public API

I provide the following public functionality:

Setup

setup/1

remove/1

do_setup/1

ensure_new/2

do_ensure_new/2

Writing

put_snapshot/2

delete_key/2

put/3

Reading

snapshot_order/1

get/2

get_keyspace/2

get_at_snapshot/2

in_snapshot/2

read_order/2

read_order_tx/2

read_at_order/3

read_at_order_tx/3

Blocking

blocking_read/2

Other

cm_tree_spec/0

Types

order_key()

@type order_key() :: Noun.t()

I designate the type of keys for Ordering table queries.

order_value()

@type order_value() :: [{any(), Noun.t(), non_neg_integer()}]

I designate the type of values of keys for Ordering table queries.

qualified_key()

@type qualified_key() :: nonempty_improper_list(any(), non_neg_integer())

I designate the type of keys for Qualified table queries.

qualified_value()

@type qualified_value() :: any()

I designate the type of values of Keys for Qualified table queries.

result(res)

@type result(res) :: {:atomic, res} | {:aborted, any()}

I am the type of a mensia table query result.

snapshot()

@type snapshot() :: {t(), [{order_key(), non_neg_integer()}]}

I designate the snapshot type for table queries.

t()

@type t() :: %Anoma.Node.Storage{
  namespace: [binary()],
  order: atom(),
  qualified: atom(),
  rm_commitments: atom(),
  topic: Anoma.Node.Router.Addr.t() | nil
}

I am the type of the Storage Engine.

I store info on the db tables used by the Node and info for appropriate read/write actions.

Fields

:qualified - The name of the Qualified table.
           Default: Anoma.Node.Storage.Qualified

:order - The name of the Ordering table.
       Default: Anoma.Node.Storage.Order

:rm_commitments - The name of the Reource Machine Commitments table.
                Default: Anoma.Node.Storage.RMCommitments

:namespace - The namespace used for writing functionality.
           Default: []

:topic - The subscription topic.
       Enforced: false

Functions

blocking_read(storage, key)

@spec blocking_read(Anoma.Node.Router.Addr.t(), qualified_key()) ::
  :error | {:ok, any()}

I provide the final part of the blocking functionality for the Worker necessary for transaction-execution.

General Usage

I expect a Storage Engine address alingside with a key. Given a key which isn't a list, I error. Given a key which is a list starting with 0, I error following Nock and Ordering semantics. Given any other list, I subscribe to a Qualified table that is attached to the given Storage and read it in the Qualified table.

If the key has a non-empty value in the table, I return {:ok, value} and unsubscribe from the table.

If the key does not have a value in the table, I wait until I receive a :write message from the table, where the key written is exactlythe key supplied. Given the message, I return the written value as {:ok, value} and unsubscribe from the table.

Anoma-Intended Semantics

My intended use in the Anoma Lifecycle is as follows:

I am called by the Worker after assigning its transaction ID to the order assigned for block-generation. The key here is given as a cell of form [n | snapshot_path] where n is a natural number representing the n-th transaction candidate - i.e. the order of the transaction the Worker is occupied with - and snapshot_path is a Nock list, i.e. an improper list of form [... | 0].

My main functionality is to make sure that the value of said transaction has been recorded suscessfully in the Qualified table in the following way:

After subscribing to the appropriate Qualified table, I first check the fed-in Nock noun. If it starts with 0, I produce :error, per standard semantics. Given another Nock cell, we then read it as a key inside our Qualified table. If the value has already been written, we return it as an {:ok, value} tuple.

Otherwise, if we read-off an empty list, i.e. if the appropriate transaction value has not been written, we wait for the message that the exact key has been written with some value in the table. Then we return the value written in an {:ok, value} tuple.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delete_key(storage, key)

@spec delete_key(Anoma.Node.Router.Addr.t(), order_key()) :: :ok

I provide the main tombstoning functionality for the Storage Engine.

My main goal is to provide the API to make a key indistinguishable from being absent in the underlying table. That is, I use put/3 to make the latest key value :absent. This way, I make it impossible to distinguish between the given key never been put into the Qualified table and being deleted using the main get/2 functionality.

do_ensure_new(storage, rocks \\ false)

@spec do_ensure_new(t(), boolean()) :: :ok | {:aborted, any()}

I ensure that the fed in storage names correspond to completely empty new tables.

I first delete all tables as specified in the Storage Engine structure, afterwards creating the tables with given names using do_setup/2.

do_setup(t, rocks \\ false)

@spec do_setup(t(), boolean()) :: :ok | {:aborted, any()}

I setup storage with the given tables from the Storage Engine structure.

I will try to create all values of storage, even if the first one fails due to already being created, we will try the others, setting them using rocksdb depending on the second argument given in the argument.

ensure_new(storage, rocks \\ false)

@spec ensure_new(Anoma.Node.Router.Addr.t(), boolean()) :: :ok

I am the function ensuring appropriate db functionality.

Given a Storage address and a boolean flag, I first use remove/1 functionality to delete specified tables of the Engine and then ask for a setup using setup/1 functionality with an appropriate flag for rocksdb copies.

get(storage, key)

@spec get(Anoma.Node.Router.Addr.t(), order_key()) ::
  :absent | {:ok, qualified_value()} | {:error, :timed_out}

I am the main reading function for the Storage functionality.

Given a storage address and key name in the Order format, I look up the latest value of such a key in the Qualified storage and return it in the standard mnesia format. Otherwise, I return :absent if no value is present (or if the key is tombstoned* in the latest instance).

See delete_key/2

get_at_snapshot(snapshot, key)

@spec get_at_snapshot(snapshot(), order_key()) :: :absent | {:ok, qualified_value()}

I provide the main functionality of rescuing latest snapshot value of a given key.

Given a {storage, snapshot} pair and a key, I search for the order of the key as captured by the snapshot using in_snapshot/2. Afterwards, I check the value of said key in said position in the provided table, returning said value in the standard mnesia format.

get_keyspace(storage, key_space)

@spec get_keyspace(Anoma.Node.Router.Addr.t(), [order_key()]) ::
  :absent | [{list(), qualified_value()}] | {:atomic, any()}

I am the keyspace reading function for the Storage Engine.

Given a storage address and a list of keys in the appropriate format, I first look up whether the keys appear in any qualified key in the Qualified table. If so, look up all apropriate latest values of all such keys and return them as a list of table query responses in a standard mnesia format. If not present, return :absent

handle_cast(arg1, _, storage)

in_snapshot(arg, key)

@spec in_snapshot(snapshot(), order_key()) :: nil | non_neg_integer()

I am the function searching for a key position inside of a given Ordering snapshot.

Given a {key, snapshot} tuple I go though the snapshot searching for the key in the format we store it inside the tables (i.e. namespaced). After which, I return the stored value, i.e. the latest Order at the point of snapshot-taking.

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  [
    qualified: atom(),
    order: atom(),
    rm_commitments: atom(),
    namespace: [binary()],
    topic: Anoma.Node.Router.Addr.t()
  ]
  | %{
      qualified: atom(),
      order: atom(),
      rm_commitments: atom(),
      namespace: [binary()],
      topic: Anoma.Node.Router.Addr.t()
    }
) :: {:ok, t()}

I am the initialization function for a Storage Engine Instance.

Pattern-Macthing Variations

init(%Storage{}) - I initialize the Engine with the given state.

init(args) - I expect a keylist/map with all the keywords matching
           the Storage Engine type fields. I then setup the tables
           and return the appropriate Engine state for startup.

put(storage, key, value)

@spec put(Anoma.Node.Router.Addr.t(), order_key(), qualified_value()) :: :ok

I provide the main writing functionality of the Storage Engine.

I am responsible for assigning the given values to key while respecting the underlying table structure.

That is, given a key and a value, I first namespace the key getting the namespace info from the appropriate storage state. Secondly, I re-write the order-value assigned to said namespaced key, incrementing its value. Finally, I write the value given to me to the Qualified table, assigning it to the namspaced key with appropriate new order attached, corresponding to the one in the Ordering table.

put_snapshot(storage, key)

@spec put_snapshot(Anoma.Node.Router.addr(), order_key()) :: :ok

I am the function responsible for writing an Ordering snapshot to our db.

Given a Storage Address and a key, I first snapshot the appropriate Ordering table - similarly to snapshot_order/1 - and then put it using the functionality underlying put/3 as a value for the supplied key.

Consult put/3 on the exact mechanism behind our writing.

read_at_order(storage, key, order)

@spec read_at_order(t(), Noun.t(), non_neg_integer()) :: [
  {atom(), qualified_key(), qualified_value()}
]

I am the read_at_order function.

I provide detailed reading functionality for the qualified table. Given a storage structure, a key, and an order, I read the Qualified mnesia table at a value which on a high-level corresponds to reading the key value at that particular order, which is, of course, represented by the order being part of the searched key at the mnesia level.

I return a list of 3-tuples with elements:

Name of Qualified Table

Key (note here "key" has different semantics than top-level API as
 noted above, it has a lot of extra info such as the order)

Value of the key

read_at_order_tx(storage, key, order)

@spec read_at_order_tx(Anoma.Node.Router.Addr.t(), Noun.t(), non_neg_integer()) ::
  result([{atom(), qualified_key(), qualified_value()}])

I am the function getting the value of a key at a particular order.

My main role is to get transactions at the order given by the Ordering Engine. Generally, the user is to inquire about the value of a key using get/2 functionality.

Given a Storage address, a key, and the order, I return the list containing the name of the storage, the key as given in the inputs, and its value at the specified order in the traditional mnesia format using :mnesia.transaction capabilities.

See read_at_order/3 for how we actually read the value from the Qualified table.

read_keyspace_order(storage, key_query)

@spec read_keyspace_order(t(), list()) :: result([[any()]])

I am the read_keyspace_order function.

My functionality is very similar to read_order/2 yet instead of using the mnesia read I use select and pick out any keys of list type in the Ordering table which might contain inside the given keylist.

I return the {:atomic, lst} tuple where the latter value is a list of lists where the head are the denamespaced keys containing our original query keylist and the tail being the order.

read_order(storage, key)

@spec read_order(t(), order_key()) :: [{atom(), Noun.t(), non_neg_integer()}]

I am the read_order function.

Given a storage structure and a key, I read the order of said key by literally asking mnesia to read the Ordering table at the key which is gotten by appending the storage namespace to the initially given key.

I return a list of three-tuples with following elements:

Ordering table name

Key

Order number

read_order_tx(storage, key)

@spec read_order_tx(Anoma.Node.Router.Addr.t(), order_key()) :: result(order_value())

I am the main function for getting orders of keys.

Given a Storage address and a key, I return the key with its specified order in the Ordering table Storage using :mnesia.transaction to get the value in standard mnesia format.

See read_order/2 for the exact mechanism behind order-reading.

remove(storage)

@spec remove(Anoma.Node.Router.Addr.t()) :: :ok

I am the storage removal function.

Given a Storage Engine address, I ask mnesia to delete all tables found inside said Engine structure.

setup(t)

@spec setup(Anoma.Node.Router.Addr.t()) :: :ok

I am the setup function for Storage.

Given a Storage address, I look for the tables inside its structure, and then try to create all mnesia tables with the given names. On success, I also create a new commitment tree.

snapshot_order(storage)

@spec snapshot_order(Anoma.Node.Router.Addr.t()) :: result(snapshot())

I provide a snapshotting functionality for the Ordering table.

Given a Storage address, I look up the Ordering table attached to the Engine and provide a snapshot of it as a list of keys with their orders returned in a mnesia result format and the attached storage structure.