Anoma.Node.Storage.Ordering (Anoma v0.11.0)

I am a simple mnesia-backed key-value store in an anoma node.

Currently we do not have a way of propagating what keys users want to store, thus we take the following approach to deterministic state reading:

We keep a next_order, which represents the next_order a transaction will have. From here the scry reads specifically [next_order, :key_space | 0] to get a map of the current keys saved for this node.

We keep a hash_to_order to cache the id => order mapping

Types

key()

@type key() :: any()

ordered_transactions()

@type ordered_transactions() :: [Anoma.Order.t()]

t()

@type t() :: %Anoma.Node.Storage.Ordering{
  hash_to_order: %{required(key()) => non_neg_integer()},
  logger: Anoma.Node.Router.Addr.t() | nil,
  next_order: non_neg_integer(),
  table: Anoma.Storage.t()
}

Functions

caller_blocking_read_id(ordering, list)

@spec caller_blocking_read_id(Anoma.Node.Router.Addr.t(), Noun.t()) ::
  :error | {:ok, any()}

Translate from ids to true order

Parameters

ordering - the ordering process

[id | subkey] -

the process identification for consistent reads, and a subkey for the storage key. This is akin to the Storage.qualified_key in Storage.blocking_read/2

Returns

returns the given key at a specific value

get_storage(ordering)

@spec get_storage(Anoma.Node.Router.Addr.t()) :: Anoma.Storage.t()

handle_call(arg1, from, state)

handle_cast(arg1, from, state)

handle_new_order(ordered_transactions, state)

@spec handle_new_order(ordered_transactions(), t()) ::
  {non_neg_integer(), %{required(key()) => non_neg_integer()}}

hard_reset(ordering, initial_snapshot)

@spec hard_reset(Anoma.Node.Router.Addr.t(), atom()) :: :ok

init(state)

new_order(ordering, ordered)

@spec new_order(Anoma.Node.Router.Addr.t(), ordered_transactions()) ::
  :error | {:ok, any()}

next_order(ordering)

@spec next_order(Anoma.Node.Router.Addr.t()) :: non_neg_integer()

reset(ordering)

@spec reset(Anoma.Node.Router.Addr.t()) :: :ok

state(ordering)

@spec state(Anoma.Node.Router.Addr.t()) :: t()

true_order(ordering, id)

@spec true_order(Anoma.Node.Router.Addr.t(), any()) :: non_neg_integer() | nil