Anoma.Node.Transport (Anoma v0.20.0)

I am the Transport Engine.

I implement the transport layer for communication between nodes. Connections between nodes are established by means of a two-step handshake protocol.

Public API

I provide the following public functionality:

start_server/2

receive_chunk/2

learn_node/3

learn_engine/3

send/3

new_connection/2

disconnect/2

Types

listen_addr()

@type listen_addr() :: {:tcp, binary(), :inet.port_number()} | {:unix, binary()}

I am a listener address type.

Options

:tcp - I provide host and port number for TCP socket.

:unix - I provide file path for Unix socket.

t()

@type t() :: %Anoma.Node.Transport{
  connection_pool: Supervisor.supervisor() | nil,
  connection_states: %{
    required(Anoma.Node.Router.addr()) =>
      Anoma.Node.Transport.ConnectionState.t()
  },
  known_engines: %{
    required(Anoma.Crypto.Id.Extern.t()) => Anoma.Crypto.Id.Extern.t()
  },
  known_nodes: MapSetMap.t(Anoma.Crypto.Id.Extern.t(), transport_addr()),
  logger: Anoma.Node.Router.addr() | nil,
  node_connections:
    MapSetMap.t(Anoma.Crypto.Id.Extern.t(), Anoma.Node.Router.addr()),
  node_internal_id: Anoma.Crypto.Id.t() | nil,
  pending_outgoing_engine_messages: %{
    required(Anoma.Crypto.Id.Extern.t()) => [binary()]
  },
  pending_outgoing_messages: %{
    required(Anoma.Crypto.Id.Extern.t()) => [binary()]
  },
  router: Anoma.Node.Router.addr() | nil,
  servers: %{required(transport_addr()) => Anoma.Node.Router.addr()},
  transport_internal_id: Anoma.Crypto.Id.t() | nil
}

I am the type of the Transport Engine.

Fields

:router - The address of the Router Engine that the Transport Engine instance serves to.

:logger - The Logger Engine address. Enforced: false.

:node_internal_id - Id of this node.

:transport_internal_id - Id of the Transport Engine.

:connection_pool - The supervisor which manages the connection pool that the TCPConnection Engine instance belongs to.

:node_connections - For every current node connection, maps the correspondent id to their address. Default: empty.

:connection_states - Maps current connections to their respective connection states. Default: empty.

:pending_outgoing_messages - For every node, to which we have not established a connection yet, store a list of messages addressed to the node. Default: empty.

:pending_outgoing_engine_messages - For every engine, whose node location is not known yet, store a list of messages addressed to the engine. Default: empty.

:servers - Stores a set of transport servers that the Transport Engine is currently running. Every transport server is mapped to an external transport address, by which messages may reach the Engine through this server.

:known_nodes - Maps every known node (indexed by its public id) to a set of transport addresses through which the node can be reached.

:known_engines - Stores a set of known engines by their public id: every engine id is mapped to a public id of a node that the engine belongs to.

transport_addr()

@type transport_addr() :: {:tcp, binary(), :inet.port_number()} | {:unix, binary()}

I am a transport address type.

Options

:tcp - I provide host and port number for TCP socket.

:unix - I provide file path for Unix socket.

transport_type()

@type transport_type() :: :tcp | :unix

I am a transport type.

Option

:tcp - TCP socket

:unix - Unix socket

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

disconnected(transport, reason)

@spec disconnected(Anoma.Node.Router.addr(), binary()) :: :ok

handle_cast(arg, from, s)

init(arg)

Callback implementation for GenServer.init/1.

learn_engine(transport, engine, node)

@spec learn_engine(
  Anoma.Node.Router.addr(),
  Anoma.Crypto.Id.Extern.t(),
  Anoma.Crypto.Id.Extern.t()
) ::
  :ok

I notify the Transport Engine that the specified engine lives on the specified node.

learn_node(transport, id, addr)

@spec learn_node(
  Anoma.Node.Router.addr(),
  Anoma.Crypto.Id.Extern.t(),
  transport_addr()
) :: :ok

I notify the Transport Engine that the node specified by the given public id can be reached through the specified transport address.

If there are pending outgoing messages addressed to the node, I try to initiate a new connection to the node.

new_connection(transport, trans_type)

@spec new_connection(Anoma.Node.Router.addr(), transport_type()) :: :ok

I notify the Transport Engine that you, a local engine, are now managing a new incoming connection of the specified type.

receive_chunk(transport, chunk)

@spec receive_chunk(Anoma.Node.Router.addr(), binary()) :: :ok

I notify the Transport Engine that the given chunk of data has been received.

I am normally called by a Transport Connection instance.

send(transport, dst, msg)

@spec send(Anoma.Node.Router.addr(), Anoma.Crypto.Id.Extern.t(), binary()) :: :ok

I send the specified message to the node with the specified id.

start_server(transport, addr)

@spec start_server(Anoma.Node.Router.addr(), listen_addr()) :: :ok

I try to start a new transport server listening on the specified address.

If a new transport server fails to start, the Transport Engine state does not change.

transport_type(addr)

@spec transport_type(transport_addr() | listen_addr()) :: transport_type()

I return transport type of the given transport address.