Anoma.Node.Transport.Router (Anoma v0.23.0)

I am the router module. I route messages to local and remote engines.

Public API

I provide the following public functionality:

start_tcp_server/0

start_tcp_client/0

Types

router_config()

@type router_config() :: map()

I am the type of the configuration to start a router process.

t()

@type t() :: %Anoma.Node.Transport.Router{config: router_config() | nil}

I am the state of the router.

My fields contain information to communicate with other engines.

Fields

:config - The configuration parameters of the node.

Functions

async_send(engine_id, engine_type, message)

@spec async_send(Anoma.Crypto.Id.t(), atom(), any()) :: :ok

I send an async message to an engine by sending it to its proxy. The proxy will then forward it to via a connection (e.g., TCP).

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

dump_register()

I am a helper function to dump the entire register. I am used for debugging purposes.

handle_lookup_engine(id, engine_type)

@spec handle_lookup_engine(Anoma.Crypto.Id.t(), atom()) ::
  {:error, :not_found} | {:ok, pid()}

I lookup the pid of a proxy engine for a given id and type.

Examples

iex> Router.lookup_engine(some_id, :router)

iex> Router.lookup_engine(some_id, :router)

handle_start_proxy_engine(id, state)

@spec handle_start_proxy_engine(Anoma.Crypto.Id.t(), t()) :: {:ok, t()}

I start a new proxy engine for a given router id.

handle_start_tcp_client(host, port, state)

@spec handle_start_tcp_client(
  :inet.socket_address() | :inet.hostname(),
  :inet.port_number(),
  t()
) :: {:ok, t()}

I start a new TCP client process that will connect to a remote TCP server.

handle_start_tcp_server(state)

@spec handle_start_tcp_server(t()) ::
  {:ok, t()} | {:error, :failed_to_start_tcp_server, t()}

I start a new TCP server process.

I use the configuration of the router to start the server on the proper port and host.

lookup_engine(engine_id, engine_type)

Given the id of a node and engine type, I lookup the address of the engine's proxy.

start_link(list)

start_proxy_engine(engine_id)

@spec start_proxy_engine(any()) :: :ok

@spec start_proxy_engine(Anoma.Crypto.Id.t()) :: :ok

Given the id of an engine, I start a proxy engine for this id. This proxy can be used to route messages to a remote node.

start_tcp_client(host, port)

@spec start_tcp_client(
  :inet.socket_address() | :inet.hostname(),
  :inet.port_number()
) :: :ok

I start a new TCP client to connect to a remote node at the given host and port.

start_tcp_server()

@spec start_tcp_server() :: :ok

I start a new TCP server process.

sync_send(engine_id, engine_type, message, timeout \\ 0)