Anoma.Storage (Anoma v0.11.0)

I am the Anoma Storage engine, I consist of two parts:

An ordering map which tells of the latest order in the qualification map

A qualification map which maps from a qualified key to the stored value

Types

A good way to view this is that for the t(), the fields for what is stored in mnesia, are simply the order* and qualified* values

type t Anoma.Storage to find them all.

Please also type t Anoma.Storage.t() to find out more about the central type

API

The important functions for this API are

setup/1

ensure_new/1

get/2

get/3

put/3

put/4

blocking_read/2

blocking_read/3

For Querying keyspace the following functions are useful

get_keyspace/2

If one wants to query the tables by hand then there are manual functions, but beware, this is an unintended way of using the API

query_key_space/2

read_order/2

read_order_tx/2

read_at_order/3

read_at_order_tx/3

write_at_order/4

write_at_order_tx/4

Please see my testing module AnomaTest.Storage to learn more on how to use me

Snapshots

One can snapshot the keys provided in the code by running the following

snapshot_order/1

put_snapshot/2

in_snapshot/2

get_at_snapshot/2

Types

order_key()

@type order_key() :: Noun.t()

The key we wish to store, also used for order lookup

order_value()

@type order_value() :: [{any(), Noun.t(), non_neg_integer()}]

qualified_key()

@type qualified_key() :: nonempty_improper_list(any(), non_neg_integer())

[non_neg_integer(), Noun.t() | non_neg_integer()]

qualified_value()

@type qualified_value() :: any()

result(res)

@type result(res) :: {:atomic, res} | {:aborted, any()}

snapshot()

@type snapshot() :: {t(), [{order_key(), non_neg_integer()}]}

t()

@type t() :: %Anoma.Storage{order: atom(), qualified: atom(), rm_commitments: atom()}

I represent the qualified and ordered data of storage

Fields

:qualified - The key value value map into storage

:order - A mapping from keys to the properly qualified keys

Functions

blocking_read(storage, key)

@spec blocking_read(t(), qualified_key()) :: :error | {:ok, any()}

calculate_order(list)

@spec calculate_order([{any(), any(), number()}]) :: number()

check_if_any_absent(latest_keys)

cm_tree_spec()

@spec cm_tree_spec() :: CommitmentTree.Spec.t()

ensure_new(storage)

@spec ensure_new(t()) :: :ok | {:aborted, any()}

get(storage, key)

@spec get(t(), order_key()) :: :absent | {:ok, qualified_value()}

get_at_snapshot(snapshot, key)

@spec get_at_snapshot(snapshot(), order_key()) :: :absent | {:ok, qualified_value()}

get_keyspace(storage, key_space)

@spec get_keyspace(t(), [any()]) ::
  :absent | [{qualified_key(), qualified_value()}] | {:atomic, any()}

in_snapshot(arg, key)

@spec in_snapshot(snapshot(), order_key()) :: nil | non_neg_integer()

put(storage, key, value)

@spec put(t(), order_key(), qualified_value()) :: result(:ok)

put_snapshot(storage, key)

@spec put_snapshot(t(), order_key()) :: result(:ok)

read_at_order(storage, key, order)

@spec read_at_order(t(), Noun.t(), non_neg_integer()) :: [
  {atom(), qualified_key(), qualified_value()}
]

read_at_order_tx(storage, key, order)

@spec read_at_order_tx(t(), Noun.t(), non_neg_integer()) ::
  result([{atom(), qualified_key(), qualified_value()}])

read_keyspace_order(storage, key_query)

@spec read_keyspace_order(t(), list()) :: result([[any()]])

Reads the given keyspace to obtain the latest orders for the keys

read_order(storage, key)

@spec read_order(t(), order_key()) :: [{atom(), Noun.t(), non_neg_integer()}]

read_order_tx(storage, key)

@spec read_order_tx(t(), order_key()) ::
  result([{atom(), Noun.t(), non_neg_integer()}])

remove(storage)

@spec remove(t()) :: :ok | {:aborted, any()}

setup(t)

@spec setup(t()) :: :ok | {:aborted, any()}

I setup storage with the given tables: t().

I will try to setup all values of storage, even if the first one fails due to already being setup, we will try the others.

snapshot_order(storage)

@spec snapshot_order(t()) :: result(snapshot())

write_at_order(storage, key, value, order)

@spec write_at_order(t(), Noun.t(), Noun.t(), non_neg_integer()) :: :ok

write_at_order_tx(storage, key, value, order)

@spec write_at_order_tx(t(), Noun.t(), Noun.t(), non_neg_integer()) :: result(:ok)