Anoma.Transaction (Anoma v0.11.0)

I represent an Anoma Transaction

I can be viewed as a wrapper over Anoma.Intent where I contain the intents used in a transaction

Types

execution()

@type execution() :: {:kv | :rm, Noun.t()}

t()

@type t() :: %Anoma.Transaction{id: Noun.t(), pid: pid(), transaction: execution()}

Functions

id(t)

@spec id(t()) :: Noun.t()

new(id, pid, transaction)

@spec new(Noun.t(), pid(), execution()) :: t()

pid(t)

@spec pid(t()) :: pid()

transaction(t)

@spec transaction(t()) :: execution()