EventBroker.Registry (Anoma v0.28.0)

I am the Registry for the PubSub system.

I am the central registry of all the topic subscriptions and filters. I am responsible for spawning filter agents, (un)subscribing to them, and keeping track of relations between them.

Types

registered_filters()

@type registered_filters() :: %{required(EventBroker.filter_spec_list()) => pid()}

t()

@type t() :: %EventBroker.Registry{
  registered_filters: registered_filters(),
  supervisor: atom()
}

I am the type of the Registry.

My main functionality is to keep track of all spawned filter actors.

Fields

:supervisor - The name of the dynamic supervisor launched on start.

:registered_filters - The map whose keys are a filter-spec dependency list and whose values are PID's of filter agents corresponding to said lists. Default: %{}

Functions

%EventBroker.Registry{}
(struct)

I am the type of the registered filters, matching a filter agent to its PID.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(args)

@spec start_link(list()) :: GenServer.on_start()