Examples.ENock (Anoma v0.28.0)

Functions

abs()

@spec abs() :: Noun.t()

abs_arm()

@spec abs_arm() :: Noun.t()

bex()

@spec bex() :: Noun.t()

bex_arm()

@spec bex_arm() :: Noun.t()

The bex arm for taking bex:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localbex => logics |= a=@ (bex a)

and then grabbing the arm of localbex.

call_split_test()

@spec call_split_test() :: any()

cmp()

@spec cmp() :: Noun.t()

cmp_arm()

@spec cmp_arm() :: Noun.t()

I represent the cmp gate call as a 2-argument gate.

Can be obtained by defining

=lcmp => logics |= [a=@s b=@s] (cmp [a b])

and computing

.* lcmp [0 2]

counter_arm()

@spec counter_arm() :: Noun.t()

The counter arm.

Availiable through counter:logics core.

counter_logic()

@spec counter_logic() :: Noun.t()

cue()

@spec cue() :: Noun.t()

cue_arm()

@spec cue_arm() :: Noun.t()

A cue arm for taking cue:anoma out of the logics core environment.

Can be gotten by defining gate locally as

=localcue => logics |= a=@ (cue a)

and then grabbing the arm of localcue.

dec()

@spec dec() :: Noun.t()

dec_arm()

@spec dec_arm() :: Noun.t()

The decrement arm in the tests core.

Availiable through use-dec:tests core.

dif()

@spec dif() :: Noun.t()

dif_arm()

@spec dif_arm() :: Noun.t()

I represent the dif gate call as a 2-argument gate.

Can be obtained by defining

=ldif => logics |= [a=@ b=@] (dif [a b])

and computing

.* ldif [0 2]

dul()

@spec dul() :: Noun.t()

dul_arm()

@spec dul_arm() :: Noun.t()

I represent the dul gate call as a 2-argument gate.

Can be obtained by defining

=ldul => logics |= [a=@s b=@] (dul [a b])

and computing

.* ldul [0 2]

factorial()

@spec factorial() :: Noun.t()

factorial_arm()

@spec factorial_arm() :: Noun.t()

I am the battery of the fib:tests gate of the anoma stadard library.

You can dump me by calling

.* fib:tests [0 2]

fra()

@spec fra() :: Noun.t()

fra_arm()

@spec fra_arm() :: Noun.t()

I represent the fra gate call as a 2-argument gate.

Can be obtained by defining

=lfra => logics |= [a=@s b=@s] (fra [a b])

and computing

.* lfra [0 2]

inc(key \\ "key")

@spec inc(Noun.t()) :: Noun.t()

incorrectly_ending()

@spec incorrectly_ending() :: :error

incorrectly_nested_noun()

@spec incorrectly_nested_noun() :: :error

incorrectly_starting()

@spec incorrectly_starting() :: :error

increment_counter_val(val)

@spec increment_counter_val(Noun.t()) :: Noun.t()

indexed_noun()

@spec indexed_noun() :: Noun.t()

jam()

@spec jam() :: Noun.t()

jam_and_cue(jam_value, cue_value)

@spec jam_and_cue(any(), any()) :: any()

jam_arm()

@spec jam_arm() :: Noun.t()

A cue arm for taking jam:anoma out of the logics core environment.

Can be gotten by defining gate locally as

=localjam => logics |= a=@ (jam a)

and then grabbing the arm of localjam.

jamming_and_cueing()

@spec jamming_and_cueing() :: :ok

lsh0()

@spec lsh0() :: Noun.t()

I evaluate lsh at block size 0 and gate-input [2 6].

lsh(0) evaluates the gate of the block door at block size 0, [6 1 2 6] replaces the sample with [2 6].

lsh1()

@spec lsh1() :: Noun.t()

I evaluate lsh at block size 1 and gate-input [2 6].

lsh(1) evaluates the gate of the block door at block size 1, [6 1 2 6] replaces the sample with [2 6].

lsh2()

@spec lsh2() :: Noun.t()

I evaluate lsh at block size 1 and gate-input [2 6].

lsh(2) evaluates the gate of the block door at block size 2, [6 1 2 6] replaces the sample with [2 6].

lsh(value)

@spec lsh(Noun.t()) :: Noun.t()

I am an lash arm in the block door.

My index inside the door can be seen by asking to dump the logic of =llsh => logics |= a=@ lsh:block

lte()

@spec lte() :: Noun.t()

lte_arm()

@spec lte_arm() :: Noun.t()

I represent the lte gate call as a 2-argument gate.

Can be obtained by defining

=llte => logics |= [a=@s b=@s] (lte [a b])

and computing

.* llte [0 2]

mat()

@spec mat() :: Noun.t()

mat_arm()

@spec mat_arm() :: Noun.t()

The mat arm for taking mat:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localmat => logics |= a (mat a)

and then grabbing the arm of locamix.

met0()

@spec met0() :: Noun.t()

I evaluate met at block size 0 and gate-input 28.

met(0) evaluates the gate of the block door at block size 0, [6 1 28] replaces the sample with 28.

met1()

@spec met1() :: Noun.t()

I evaluate met at block size 1 and gate-input 28.

met(1) evaluates the gate of the block door at block size 1, [6 1 28] replaces the sample with 28.

met2()

@spec met2() :: Noun.t()

I evaluate met at block size 2 and gate-input 28.

met(2) evaluates the gate of the block door at block size 2, [6 1 28] replaces the sample with 28.

met(value)

@spec met(Noun.t()) :: Noun.t()

I am an lash arm in the block door.

My index inside the door can be seen by asking to dump the logic of =lmet => logics |= a=@ met:block

mix()

@spec mix() :: Noun.t()

mix_arm()

@spec mix_arm() :: Noun.t()

The mix arm for taking mix:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localmix => logics |= [a=@ b=@] (mix [a b])

and then grabbing the arm of locamix.

nesting_noun()

@spec nesting_noun() :: Noun.t()

new()

@spec new() :: Noun.t()

new_arm()

@spec new_arm() :: Noun.t()

I represent the new gate call as a 2-argument gate.

Can be obtained by defining

=lnew => logics |= [a=? b=@] (new [a b])

and computing

.* lnew [0 2]

og_arm()

@spec og_arm() :: Noun.t()

The gate representing an og core creation with a specified seed.

Can be gotten by defining

=l => logics |= [seed=@] ~(. og seed)

and getting it's arm with [0 2]

og_call(seed)

@spec og_call(non_neg_integer()) :: :error | {:ok, Noun.t()}

old()

@spec old() :: Noun.t()

old_arm()

@spec old_arm() :: Noun.t()

I represent the old gate call as a 2-argument gate.

Can be obtained by defining

=lold => logics |= [a=@s] (old a)

and computing

.* lold [0 2]

one_two()

@spec one_two() :: Noun.t()

pro()

@spec pro() :: Noun.t()

pro_arm()

@spec pro_arm() :: Noun.t()

I represent the pro gate call as a 2-argument gate.

Can be obtained by defining

=lpro => logics |= [a=@s b=@s] (pro [a b])

and computing

.* lpro [0 2]

rad_arm()

@spec rad_arm() :: Noun.t()

I represent the rad gate call as a 2-argument gate.

Can be gotten by defining

=lrad => logics |= [a=@ b=@] (~(rad og a) b)

rad_call(seed, range)

@spec rad_call(any(), non_neg_integer()) :: {:ok, Noun.t()}

I am function calling the rad gate of the og door with specified seed and range

rad_tests()

@spec rad_tests() :: {:ok, Noun.t()}

rads_arm()

@spec rads_arm() :: Noun.t()

I represent the rads gate call as a 2-argument gate.

Can be gotten by defining

=lrad => logics |= [a=@ b=@] (~(rads og a) b)

rads_call(seed, range)

@spec rads_call(non_neg_integer(), non_neg_integer()) :: :error | {:ok, Noun.t()}

I am function calling the rads gate of the og door with specified seed and range

rads_tests()

@spec rads_tests() :: :ok

raw_27_4()

@spec raw_27_4() :: {:ok, Noun.t()}

raw_arm()

@spec raw_arm() :: Noun.t()

I represent the raw gate call as a 2-argument gate.

Can be gotten by defining

=lraw => logics |= [a=@ b=@] (~(raw og a) b)

raw_call(seed, width)

@spec raw_call(Noun.t(), Noun.t()) :: {:ok, Noun.t()}

I am function calling the raw gate of the og door with specified seed and bitwidth.

raws_arm()

@spec raws_arm() :: Noun.t()

I represent the raws gate call as a 2-argument gate.

Can be gotten by defining

=lraw => logics |= [a=@ b=@] (~(raws og a) b)

raws_call(seed, width)

@spec raws_call(Noun.t(), Noun.t()) :: {:ok, Noun.t()}

I am function calling the raws gate of the og door with specified seed and bitwidth.

raws_test()

@spec raws_test() :: :ok

raws_with_core()

@spec raws_with_core() :: Noun.t()

I represent a raws gate with a specified instantiated og core given as an extra argument.

Can be gotten by defining locally

=l => logics |= [rng=_og width=@] (raws:rng width)

and grabbing the arm with [0 2]

raws_with_core_call(core, width)

@spec raws_with_core_call(non_neg_integer(), non_neg_integer()) ::
  :error | {:ok, Noun.t()}

raws_with_out_core_test()

@spec raws_with_out_core_test() :: any()

rem()

@spec rem() :: Noun.t()

rem_arm()

@spec rem_arm() :: Noun.t()

I represent the rem gate call as a 2-argument gate.

Can be obtained by defining

=lrem => logics |= [a=@s b=@s] (rem [a b])

and computing

.* lrem [0 2]

replacing_terms()

@spec replacing_terms() :: Noun.t()

rsh0()

@spec rsh0() :: Noun.t()

I evaluate rsh at block size 0 and gate-input [2 40].

rsh(0) evaluates the gate of the block door at block size 0, [6 1 2 40] replaces the sample with [2 40].

rsh1()

@spec rsh1() :: Noun.t()

I evaluate rsh at block size 1 and gate-input [2 40].

rsh(1) evaluates the gate of the block door at block size 1, [6 1 2 40] replaces the sample with [2 40].

rsh2()

@spec rsh2() :: Noun.t()

I evaluate rsh at block size 2 and gate-input [2 40].

rsh(2) evaluates the gate of the block door at block size 2, [6 1 1 40] replaces the sample with [1 40].

rsh(value)

@spec rsh(Noun.t()) :: Noun.t()

I am an lash arm in the block door.

My index inside the door can be seen by asking to dump the logic of =rsh => logics |= a=@ rsh:block

shax()

@spec shax() :: Noun.t()

shax_arm()

@spec shax_arm() :: Noun.t()

The shax arm for taking shax:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localshax => logics |= a=@ (shax a)

and then grabbing the arm of localshax.

sign()

@spec sign() :: Noun.t()

sign_arm()

@spec sign_arm() :: Noun.t()

The sign arm for taking sign:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localsign => logics |= [a=@ b=@] (sign [a b])

and then grabbing the arm of localsign.

sign_detatched()

@spec sign_detatched() :: Noun.t()

sign_detatched_arm()

@spec sign_detatched_arm() :: Noun.t()

The sign-detatched arm for taking sign-detached:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localsigndetached => logics |= [a=@ b=@] (sign-detached [a b])

and then grabbing the arm of localsighdetached.

split_arm()

@spec split_arm() :: Noun.t()

I represent a split gate call given an og core with seed.

Can be gotten by defining locally

=l => logics |= [rng=_og] split:rng

and grabbing the arm with [0 2]

split_call(core)

@spec split_call(Noun.t()) :: :error | {:ok, Noun.t()}

sum()

@spec sum() :: Noun.t()

sum_arm()

@spec sum_arm() :: Noun.t()

I represent the sum gate call as a 2-argument gate.

Can be obtained by defining

=lsum => logics |= [a=@s b=@s] (sum [a b])

and computing

.* lsum [0 2]

sun()

@spec sun() :: Noun.t()

sun_arm()

@spec sun_arm() :: Noun.t()

syn()

@spec syn() :: Noun.t()

syn_arm()

@spec syn_arm() :: Noun.t()

transparent_core(tx_noun \\ Examples.ETransparent.ETransaction.swap_from_actions() |> Noun.Nounable.to_noun())

@spec transparent_core(Noun.t()) :: Noun.t()

trivial_swap()

@spec trivial_swap() :: Noun.t()

trivial_swap_no_eph()

@spec trivial_swap_no_eph() :: Noun.t()

uend0()

@spec uend0() :: Noun.t()

I evaluate uend at block size 0 and gate-input [5 80].

uend(0) evaluates the gate of the block door at block size 0, [6 1 5 80] replaces the sample with [5 80].

uend1()

@spec uend1() :: Noun.t()

I evaluate uend at block size 1 and gate-input [3 80] and [4 80].

uend(1) evaluates the gate of the block door at block size 1, [6 1 3 80] replaces the sample with [3 80], [6 1 4 80] replaces the sample with [3 80]

uend(value)

@spec uend(Noun.t()) :: Noun.t()

I am an lash arm in the block door.

My index inside the door can be seen by asking to dump the logic of =luend => logics |= a=@ luend:block

verify()

@spec verify() :: Noun.t()

verify_arm()

@spec verify_arm() :: Noun.t()

The verify arm for taking verify:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localverify => logics |= [a=@ b=@] (verify [a b])

and then grabbing the arm of localverify.

verify_detatched()

@spec verify_detatched() :: Noun.t()

verify_detatched_arm()

@spec verify_detatched_arm() :: Noun.t()

The verify-detatched arm for taking verify-detached:anoma from the logics core environment.

Can be gotten by defining gate locally as:

=localverifydetached => logics |= [a=@ b=@ c=@] (verify-detached [a b])

and then grabbing the arm of localverifydetached.

zero(key \\ "key")

@spec zero(Noun.t()) :: Noun.t()

zero_counter(val)

@spec zero_counter(Noun.t()) :: Noun.t()

zero_delta_logic()

@spec zero_delta_logic() :: Noun.t()