Examples.ENode (Anoma v0.17.0)

I give examples of the ENode.

Currently I make networks in Phase_1 of the example. Meaning that network configurations I provide, are not cloned. The given storage names will be written over for any shared set of storage.

When phase_2 of the changes come in, the signature and examples of this module in particular will likely change quite a bit to return persistent networks that other examples can use non-destructively. And thus every example can simply add on to another. However since we don't have this examples, are expected to clean out the state before running.

Functions

base_snapshot_path()

@spec base_snapshot_path() :: Noun.t()

fresh_full_node(storage, name, configuration \\ nil)

@spec fresh_full_node(
  Anoma.Node.Storage.t(),
  atom(),
  Anoma.Configuration.configuration_map() | nil
) :: Anoma.Node.t()

We give an example of the full node!

intent(node)

@spec intent(Anoma.Node.t()) ::
  {Anoma.Node.Router.addr(), Anoma.Node.Router.addr(), Anoma.Node.t()}

simple_ordering(storage)

@spec simple_ordering(Anoma.Node.Storage.t()) :: Anoma.Node.t()

Simple node with the minimal dependencies for Storage

simple_router()

@spec simple_router() :: Anoma.Node.t()

I am simply a node with just a router and transport

simple_storage(storage)

@spec simple_storage(Anoma.Node.Storage.t()) :: Anoma.Node.t()

Minimal node, with storage

simple_storage_topic(storage)

@spec simple_storage_topic(Anoma.Node.Storage.t()) :: Anoma.Node.t()

Minimal node, with storage and a topic

solver(node)

@spec solver(Anoma.Node.t()) ::
  {Anoma.Node.Router.addr(), Anoma.Node.Router.addr(), Anoma.Node.Router.addr(),
   Anoma.Node.t()}

zero_clock()

I am just the clock engine with router support. My time starts at 0