Examples.ENode.EStorage (Anoma v0.19.0)

Functions

anode(arg \\ "none")

@spec anode(Anoma.Symbol.s()) :: Anoma.Node.t()

august_namespace()

august_node(storage_name \\ "august_node")

@spec august_node(Anoma.Symbol.s()) :: Anoma.Node.t()

I am the august storage, see my Keys for details on my storage setup

Keys:

miki_key/0 - lucky_value/0

isuzumi_key/0 - devil_value/0

Examples.ENode.base_snapshot_path/0 - snapshot of both

august_node_proper(storage_name \\ "august_node_proper")

@spec august_node_proper(Anoma.Symbol.s()) ::
  {Anoma.Node.t(), Anoma.Identity.Backend.Memory.t(),
   Anoma.Identity.Manager.instance(), Anoma.Crypto.Id.Extern.t()}

I am like august_node/1. Except I sign the namespace reserved

Further besides the main space being owned by memory backed space, we have ECrypto.londo/0 owning the water. Thankfully we aren't on Narn.

august_replaced_keyspace(storage_name \\ "august_replaced")

@spec august_replaced_keyspace(Anoma.Symbol.s()) :: Anoma.Node.t()

I am the replaced keyspace node for august.

I use august_node and then delete the associated miki value. This way when we print keyspaces, we see all the previously availiable key-value pairs sans the miki key.

august_space()

august_subnamespace(storage)

@spec august_subnamespace(Anoma.Node.Router.addr()) :: Anoma.Identity.Name.t()

bertha_speaks_for_all(storage_name \\ "bertha_speaks")

@spec bertha_speaks_for_all(Anoma.Symbol.s()) :: Anoma.Node.t()

blocking_for_put(storage_name \\ "blocking_for_put")

@spec blocking_for_put(Anoma.Symbol.s()) :: Anoma.Node.t()

deleting_nothing_works_fine(stg_name \\ "deleting_nothing_works_fine")

@spec deleting_nothing_works_fine(Anoma.Symbol.s()) :: Anoma.Node.t()

deleting_the_put(storage_name \\ "deleting_the_put")

@spec deleting_the_put(Anoma.Symbol.s()) :: Anoma.Node.t()

devil_key()

devil_value()

empty_storage(storage_name \\ "empty_storage")

@spec empty_storage(Anoma.Symbol.s()) :: Anoma.Node.t()

isuzumi_key()

londo_speaks_for_alice(storage_name \\ "londo_speaks")

@spec londo_speaks_for_alice(Anoma.Symbol.s()) :: Anoma.Node.t()

lucky_key()

lucky_value()

miki_key()

put_storage(storage_name \\ "put_storage")

@spec put_storage(Anoma.Symbol.s()) :: Anoma.Node.t()

random_id()

raw_storage(arg \\ "none")

@spec raw_storage(Anoma.Symbol.s()) :: Anoma.Node.Storage.t()

reserved_august(storage_name \\ "reserved_august")

@spec reserved_august(Anoma.Symbol.s()) ::
  {Anoma.Node.t(), Anoma.Identity.Backend.Memory.t(),
   Anoma.Identity.Manager.instance(), Anoma.Crypto.Id.Extern.t()}

snapshot_again(storage_name \\ "snapshot_again")

@spec snapshot_again(Anoma.Symbol.s()) :: Anoma.Node.t()

snapshot_point()

snapshot_then_put(storage_name \\ "snapshot_then_put")

@spec snapshot_then_put(Anoma.Symbol.s()) :: Anoma.Node.t()

I have Examples.Enock.one_two/2 in storage twice and a snapshot!

The snapshot only snapshots the first key though, keep that in mind.

Keys:

Examples.ENock.one_two/0 - latest value is lucky_value/0

Examples.ENode.base_snapshot_path/0 - snapshot with Examples.ENock.one_two/0 being 1

sub(anode)

@spec sub(Anoma.Node.t()) :: any()

unsub(anode)

@spec unsub(Anoma.Node.t()) :: any()