Livebook (Anoma v0.11.0)

I generate out extra information for Livebook

My main purpose is to generate out the TOC for each livebook document we have.

to do this please run toc_toplevel/0

API

toc_toplevel/0

get_all_livemd_documents/0

example_toc/0

Functions

add_heading_num(documents)

@spec add_heading_num([String.t()]) :: [
  {String.t(), non_neg_integer(), non_neg_integer()}
]

change_header(markdown, header_level, start_header, new_text)

I replace the header with the given TOC

Example

markdown_text = "## Intro ...

Index

text here ## Conclusion All good"

Livebook.change_header(markdown_text, "##", "Index", "New Content") |> IO.puts

 ## Intro ...
 ## Index
 New Content
 ## Conclusion
 All good
:ok

count_depth(path)

@spec count_depth(Path.t()) :: non_neg_integer()

dir_from_path(dir)

@spec dir_from_path(String.t()) :: [String.t()]

example_toc()

@spec example_toc() :: :ok

I provide an example of what a TOC looks like

generate_heading(path, depth, numbering, from_depth \\ 1)

@spec generate_heading(
  String.t(),
  non_neg_integer(),
  non_neg_integer(),
  non_neg_integer()
) :: String.t()

generate_TOC(documents, from_depth \\ 0)

@spec generate_TOC([Path.t()], non_neg_integer()) :: String.t()

Generates out a TOC, given a series of nested documents

We take a path, and a place where we should be calculating the TOC from.

Example

get_all_livemd_documents()

@spec get_all_livemd_documents() :: [Path.t()]

I get out all live view docs

get_livemd_documents(dir)

@spec get_livemd_documents(Path.t()) :: [Path.t()]

Gets all livemd documents in a sorted list given a path.

inject_TOC(toc, path)

@spec inject_TOC(String.t(), Path.t()) :: :ok

toc_toplevel()

I generate out the TOC for all liveview docs