Anoma.Dump (Anoma v0.11.0)

I provide an interface to dump current state and load appropriate external files to launch them as Anoma nodes.

You can also use me to dump info such as current states and tables in a readable map format as well as get info stored in external files in binary format.

Dumping API

I give access to following public dumping functionality:

dump/2

get_all/1

get_state/1

get_tables/1

Loading API

I give access to following public loading functionality

launch/2

load/1

Types

clock_eng()

@type clock_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Clock.t()}

ex_eng()

@type ex_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Executor.t()}

log_eng()

@type log_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Logger.t()}

mem_eng()

@type mem_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Mempool.t()}

ord_eng()

@type ord_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Storage.Ordering.t()}

ping_eng()

@type ping_eng() :: {Anoma.Crypto.Id.Extern.t(), Anoma.Node.Pinger.t()}

stores()

@type stores() :: {Anoma.Storage.t(), atom()}

Functions

dump(name, node)

@spec dump(String.t(), atom()) :: {:ok, :ok} | {:error, any()}

I dump the current state with storage. I accept a string as a name, so that the resulting file will be created as name.txt and then a node whose info presented as a map I dump as a binary.

The map typing can be seen in get_all

get_all(node)

@spec get_all(atom()) :: %{
  router: Anoma.Crypto.Id.Extern.t(),
  mempool_topic: Anoma.Crypto.Id.Extern.t(),
  executor_topic: Anoma.Crypto.Id.Extern.t(),
  logger: log_eng(),
  clock: clock_eng(),
  ordering: ord_eng(),
  mempool: mem_eng(),
  pinger: ping_eng(),
  executor: ex_eng(),
  storage: stores(),
  qualified: list(),
  order: list(),
  block_storage: list()
}

I get all the info on the node tables and engines in order:

router

logger

clock

ordering

mempool

pinger

executor

table names

qualified

order

block_storage And turn the info into a tuple

get_state(node)

@spec get_state(atom()) :: %{
  router: Anoma.Crypto.Id.Extern.t(),
  mempool_topic: Anoma.Crypto.Id.Extern.t(),
  executor_topic: Anoma.Crypto.Id.Extern.t(),
  logger: log_eng(),
  clock: clock_eng(),
  ordering: ord_eng(),
  mempool: mem_eng(),
  pinger: ping_eng(),
  executor: ex_eng()
}

I get the engine states in order:

router

mempool topic

executor topic

logger

clock

ordering

mempool

pinger

executor

get_tables(node)

@spec get_tables(atom()) :: %{
  storage: stores(),
  qualified: list(),
  order: list(),
  block_storage: list()
}

I get the node tables in order:

storage (names)

qualified

order

block_storage

launch(file, name)

@spec launch(String.t(), atom()) ::
  {:ok,
   %Anoma.Node{
     clock: term(),
     executor: term(),
     executor_topic: term(),
     logger: term(),
     mempool: term(),
     mempool_topic: term(),
     ordering: term(),
     pinger: term(),
     router: term()
   }}
  | any()

I launch a node given a file containing a binary version of an 12-tuple with appropriate info in the following order:

router id

mempool topic id

executor topic id

logger

clock

ordering

mempool

pinger

executor

storage names

qualified

order

block_storage

All engines have info on their states and id's so that checkpointing the system will keep all adresses used in the previous session. Note that I ensure that the apporpriate tables are new.

Check whether your transactions have had an assigned worker. If not, relaunch them directly.

load(name)

@spec load(String.t()) :: any()

I read the given file which I assume contains binary info and convert it to an Elixir term.

As the dumped state may have extra atoms not present in the session, I currently allow for atom creation in the loaded term.