Anoma.Node (Anoma v0.11.0)

I act as a registry for Anoma Nodes

There are two ways to launch a Node. Either with minimal arguments or having the full specification for all the engines, including their external ID's.

All inputs should come in format of a list of atom-value 2-tuples, {:settings, map()} specifying the info on all relevant engines and storages possibly including their tables presented in a dumped format, {:name, atom()} specifying the name of the node, and {:new_storage, bool()} specifying whether the table names, if supplied, need to be set-up without any deletions.

For more info on format check start_min/1

Minimal Arguments

name - name for this process

snapshot_path : [atom() | 0]

A snapshot location for the service (used in the worker)

storage : Anoma.Storage.t() - The Storage tables to use

block_storage - a location to store the blocks produced

Created Tables

storage.qualified

storage.order

block_storage

Types

t()

@type t() :: %Anoma.Node{
  clock: Anoma.Node.Router.addr(),
  executor: Anoma.Node.Router.addr(),
  executor_topic: Anoma.Node.Router.addr(),
  logger: Anoma.Node.Router.addr(),
  mempool: Anoma.Node.Router.addr(),
  mempool_topic: Anoma.Node.Router.addr(),
  ordering: Anoma.Node.Router.addr(),
  pinger: Anoma.Node.Router.addr(),
  router: Anoma.Node.Router.addr()
}

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(args)

Callback implementation for GenServer.init/1.

start_link(args)

I assume I am fed a list with atom-value 2-tuples :new_storage, :name, :settings. If :new_storage has value true, I need a snapshot_path: key in the settings.

I ensure that the storages are deleted and created. Afterwards, if the storage is truly new, I put the snapshot in the ordering. Otherwise, I simply repopulate the tables from a supplied list by the settings. Check Anoma.Dump for format descriptions.

start_min(args)

Given minimal arguments, I create appropriate setup for the :settings argument for Node initialization.

state(nodes)