Anoma.Node.Executor (Anoma v0.17.0)

I am the Executor Engine.

My main responsibility is to be the intemediary between the Mempool and transaction spawning by launching the Worker Engines provided the pending transaction code.

API

My public facing API is

snapshot/1

kill_transactions/2

fire_new_transaction/4

Types

t()

@type t() :: %Anoma.Node.Executor{
  ambiant_env: Nock.t() | nil,
  logger: Anoma.Node.Router.Addr.t() | nil,
  router: Anoma.Node.Router.Addr.t() | nil,
  topic: Anoma.Node.Router.Addr.t() | nil,
  workers: [Anoma.Node.Router.Addr.t()]
}

I am the type of the Executor Engine.

Fields

:router - The Router Engine address used by the Executor.

:topic - The topic for broadcasting.

:ambiant_env - The default environment to be fed to spawned workers.

:workers - A list of worker addresses spawned.
         Default: []

:logger - The Logger Engine address.
        Enforced: false

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

fire_new_transaction(executor, id, gate, env \\ nil)

@spec fire_new_transaction(
  Anoma.Node.Router.Addr.t(),
  non_neg_integer(),
  Noun.t(),
  Nock.t() | nil
) :: :ok

I am a function spawning a new transaction.

I do so by spawning a new engine with the info provided. If the environment is left unspecified I launch the Worker with the given ambient environment stored in the state.

handle_cast(arg, from, state)

init(state)

@spec init(t()) :: {:ok, t()}

@spec init(
  {Anoma.Node.Router.Addr.t(), Nock.t(), Anoma.Node.Router.Addr.t(),
   Anoma.Node.Router.Addr.t()}
) :: {:ok, t()}

### Pattern-Matching Variations

init(%Executor{}) - I initialize the Engine with the given state.

init({router, env, topic, logger}) - I expect a tuple with needed
                                   arguments to launch the executor.

kill_transactions(executor, transactions)

@spec kill_transactions(Anoma.Node.Router.Addr.t(), [Anoma.Transaction.t()]) :: :ok

I am a function killing transactions.

I currently do not filter through which transactions to kill. I just go through the list of all Workers stored in my state and kill their processes one by one. The returned state will have a nil list for the current worker list.

snapshot(executor)

@spec snapshot(Anoma.Node.Router.Addr.t()) :: atom()

I return the snapshot path.