Anoma.Node.Logging (Anoma v0.26.0)

Replay manager with logger functionality

Types

flag()

@type flag() :: :info | :debug | :error

t()

@type t() :: %Anoma.Node.Logging{node_id: String.t() | nil, table: atom()}

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(args)

@spec init(any()) :: {:ok, t()}

Callback implementation for GenServer.init/1.

init_table(table, rocks)

@spec init_table(atom(), bool()) :: {:atomic, :ok}

log_event(node_id, flag, msg)

logging_filter()

replay_args(list)

@spec replay_args(height: integer(), mempool: list()) :: [
  mempool: list(),
  ordering: list(),
  storage: list()
]

replay_setup(event_table, block_table)

@spec replay_setup(atom(), atom()) :: [height: integer(), mempool: list()]

The first step in the replay process.

All tables here are the original ones from which we get info.

This gives us the height, the round, as well as the replay Mempool struct

replay_table_clone(values_table, updates_table, node_id)

Copies the contents of the storage tables into mock Replay tables

restart_with_replay(node_id)

@spec restart_with_replay(String.t()) :: DynamicSupervisor.on_start_child()

I am the function to be played on restarts of the Anoma node with known ID.

I sync the event table with the blocks submitted and then launch a mock node with a new ID with replay arguments.

If the replay succeeds, I reproduce it on the node with the ID provided. Otherwise, the only initialization information I reproduce are heights and rounds for the Transaction subsystem.

start_link(args)

table_name(node_id)

try_launch(mock_id, replay_args)

@spec try_launch(String.t(), any()) :: :ok | :error