Anoma.Node.Transport (Anoma v0.15.0)

Types

listen_addr()

@type listen_addr() :: {:tcp, binary(), non_neg_integer()} | {:unix, binary()}

t()

@type t() :: %Anoma.Node.Transport{
  connection_pool: Supervisor.supervisor() | nil,
  connection_states: %{
    required(Anoma.Node.Router.addr()) =>
      Anoma.Node.Transport.ConnectionState.t()
  },
  known_engines: %{
    required(Anoma.Crypto.Id.Extern.t()) => Anoma.Crypto.Id.Extern.t()
  },
  known_nodes: MapSetMap.t(Anoma.Crypto.Id.Extern.t(), transport_addr()),
  logger: Anoma.Node.Router.addr() | nil,
  node_connections:
    MapSetMap.t(Anoma.Crypto.Id.Extern.t(), Anoma.Node.Router.addr()),
  node_internal_id: Anoma.Crypto.Id.t() | nil,
  pending_outgoing_engine_messages: %{
    required(Anoma.Crypto.Id.Extern.t()) => [binary()]
  },
  pending_outgoing_messages: %{
    required(Anoma.Crypto.Id.Extern.t()) => [binary()]
  },
  router: Anoma.Node.Router.addr() | nil,
  servers: %{required(transport_addr()) => Anoma.Node.Router.addr()},
  transport_internal_id: Anoma.Crypto.Id.t() | nil
}

transport_addr()

@type transport_addr() :: {:tcp, binary(), non_neg_integer()} | {:unix, binary()}

transport_type()

@type transport_type() :: :tcp | :unix

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

disconnected(transport, reason)

@spec disconnected(Anoma.Node.Router.addr(), binary()) :: :ok

handle_cast(arg, from, s)

handle_recv_chunk(s, from, data)

init(arg)

Callback implementation for GenServer.init/1.

learn_engine(transport, engine, node)

@spec learn_engine(
  Anoma.Node.Router.addr(),
  Anoma.Crypto.Id.Extern.t(),
  Anoma.Crypto.Id.Extern.t()
) ::
  :ok

Notify the transport that the specified engine lives on the specified node.

learn_node(transport, id, addr)

@spec learn_node(
  Anoma.Node.Router.addr(),
  Anoma.Crypto.Id.Extern.t(),
  transport_addr()
) :: :ok

Notify the transport that the specified node can be reached by the specified transport address.

new_connection(transport, trans_type)

@spec new_connection(Anoma.Node.Router.addr(), transport_type()) :: :ok

Notify the transport that you, a local engine, are now managing a new incoming connection of the specified type.

receive_chunk(transport, chunk)

@spec receive_chunk(Anoma.Node.Router.addr(), binary()) :: :ok

You are a connection, and want to notify the transport that you have received a chunk of data.

send(transport, dst, msg)

@spec send(Anoma.Node.Router.addr(), Anoma.Crypto.Id.Extern.t(), binary()) :: :ok

Attempt to send the specified message to the node with the specified id, somehow.

start_server(transport, addr)

@spec start_server(Anoma.Node.Router.addr(), listen_addr()) :: :ok

Attempt to start a new transport server listening on the specified address.

transport_type(addr)

@spec transport_type(transport_addr()) :: transport_type()